### 2021-09-03 #3691 @whedon
[PRE REVIEW]: Nextclade: clade assignment, mutation calling and quality control for viral genomes
**Submitting author:** @rneher (<a href="http://orcid.org/0000-0003-2525-1407">Richard Neher</a>)
**Repository:** <a href="https://github.com/nextstrain/nextclade" target ="_blank">https://github.com/nextstrain/nextclade</a>
**Version:** v1.3.0
