### 2021-09-03 #3691 @kyleniemeyer
Hi @rneher, thanks for your interest in JOSS! Unfortunately, due to low editor availability in this area, I need to put your submission on our waitlist. Once an editor is able to handle this one, we'll get it assigned and moving forward.
