### 2021-09-06 153b7ab
fix: handle some of the edge cases for compensated indels
