+++ packages/nextalign/tests/frameshiftDetector.test.cpp
@@ -467,74 +467,72 @@ TEST_F(DetectFrameShifts, DetectsCompensatedShiftDueToInsertionAndDeletion) {
   EXPECT_EQ(actual, expected);
 }
 
+TEST_F(DetectFrameShifts, DetectsCompensatedShiftDueTo2xInsertionsAnd2xDeletions) {
+  std::stringstream input;
 
-//TEST_F(DetectFrameShifts, DetectsCompensatedShiftDueTo2xInsertionsAnd2xDeletions) {
-//  std::stringstream input;
-//
-//  // clang-format off
-//  //                                                   10           20        30           40        50
-//  //                                                   |            |         |            |         |
-//  //                                         012345678901234567   89012345678901   2345678901234567890
-//  const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCT" "--AGATAACAGAAC" "ATTCTTGGAATGCTGATC" );
-//  const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCT" "ATAGATAACAGA--" "ATTCTTGGAATGCTGATC" );
-//  // frame                                   000000000000000000   12222222222210   000000000000000000
-//  // frame shift ranges                                             xxxxxxxxxx
-//  //                                                                ^         ^
-//  //                                                                20        30
-//  // clang-format on
-//
-//  const auto result = detectFrameShifts(refAln, qryAln);
-//  const auto& actual = result.frameShifts;
-//
-//  const std::vector<FrameShiftRange> expected = {
-//    FrameShiftRange{.begin = 20, .end = 30},
-//  };
-//
-//  EXPECT_EQ(actual, expected);
-//}
-//
-//
-//TEST_F(DetectFrameShifts, DetectsUncompensatedShiftDueTo2xInsertionsAnd1xDeletion) {
-//  std::stringstream input;
-//
-//  // clang-format off
-//  //                                                   10           20        30           40        50
-//  //                                                   |            |         |            |         |
-//  //                                         012345678901234567   890123456789012345678901234567890
-//  const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCT" "--AGATAACAGAACATTCTTGGAATGCTGATC" );
-//  const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCT" "ATAGATAACAGAA-ATTCTTGGAATGCTGATC" );
-//  // frame                                   000000000000000000   122222222222111111111111111111111
-//  // frame shift ranges                                             xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-//  //                                                              ^                               ^
-//  //                                                             18                               50
-//  // clang-format on
-//
-//  const auto result = detectFrameShifts(refAln, qryAln);
-//  const auto& actual = result.frameShifts;
-//
-//  const std::vector<FrameShiftRange> expected = {
-//    FrameShiftRange{.begin = 18, .end = 50},
-//  };
-//
-//  EXPECT_EQ(actual, expected);
-//}
+  // clang-format off
+  //                                                   10        20        30        40        50
+  //                                                   |         |         |         |         |
+  //                                         012345678901234567890123456789012345678901234567890
+  const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCT--AGATAACAGAACATTCTTGGAATGCTGATC" );
+  const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCTATAGATAACAGA--ATTCTTGGAATGCTGATC" );
[…]
