+++ packages/nextalign/tests/frameshiftDetector.test.cpp
@@ -54,7 +54,7 @@ TEST_F(DetectFrameShifts, DetectsNoShiftsIf3xDelsAdjacent) {
   EXPECT_EQ(actual, expected);
 }
 
-TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletaionsAreModulo3) {
+TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3) {
   std::stringstream input;
 
   // clang-format off
@@ -70,7 +70,7 @@ TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletaionsAreModulo3) {
   EXPECT_EQ(actual, expected);
 }
 
-TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletaionsAreModulo3Adjacent) {
+TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3Adjacent) {
   std::stringstream input;
 
   // clang-format off
@@ -126,6 +126,39 @@ TEST_F(DetectFrameShifts, DetectsNoShiftsIfInsertionsAreModulo3Adjacent) {
   // clang-format off
   const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCG---CTATAGATAACAGAACA---TTGGAATGCTGATC" );
   const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCTATAGATAACAGAACATTCTTGGAATGCTGATC" );
+  // clang-format on
+
+  const auto result = detectFrameShifts(refAln, qryAln);
+  const auto& actual = result.frameShifts;
+
+  const std::vector<FrameShiftRange> expected = {};
+
+  EXPECT_EQ(actual, expected);
+}
+
+TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletionsAndInsertionsAreModulo3AndAdjacent) {
+  std::stringstream input;
+
+  // clang-format off
+  const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCG---CTATAGATAACAGAACATTCTTGGAATGCTGATC" );
+  const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCCGTGG---TAGATAACAGAACATTCTTGGAATGCTGATC" );
+  // clang-format on
+
+  const auto result = detectFrameShifts(refAln, qryAln);
+  const auto& actual = result.frameShifts;
+
+  const std::vector<FrameShiftRange> expected = {};
+
+  EXPECT_EQ(actual, expected);
+}
+
+
+TEST_F(DetectFrameShifts, DetectsNoShiftsIfInsertionsAndDeletionsAreModulo3AndAdjacent) {
+  std::stringstream input;
+
+  // clang-format off
+  const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCG---CTATAGATAACAGAACATTCTTGGAATGCTGATC" );
+  const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCCGTGG---TAGATAACAGAACATTCTTGGAATGCTGATC" );
   // clang-format on
 
   const auto result = detectFrameShifts(refAln, qryAln);
@@ -240,7 +273,33 @@ TEST_F(DetectFrameShifts, DetectsCompensatedShiftDueTo3Deletions) {
 }
 
 
-TEST_F(DetectFrameShifts, DetectsUncompensatedShiftDueTo2xAdjacentDeletionsLeading) {
+TEST_F(DetectFrameShifts, DetectsUncompensatedShiftDueToDeletionLeading) {
+  std::stringstream input;
+
+  // clang-format off
+  //                                                   10        20        30        40        50
+  //                                                   |         |         |         |         |
[…]
