### 2021-09-06 6ee1f10
fix: handle more frame shift edge cases
