### 2021-09-06 2c77433
refactor: add shift, instead of subtracting
So `that` frame values match the annotation comments in uni tests. Does not influence the algorithm behavior so far.
