### 2021-09-06 a8ee2ec
test: prettify frame shift detector tests and add empty case
