+++ packages/nextalign/tests/frameshiftDetector.test.cpp
@@ -17,13 +17,10 @@ protected:
   }
 };
 
-TEST_F(DetectFrameShifts, DetectsNoShiftsIfNoIndels) {
-  std::stringstream input;
-
+TEST_F(DetectFrameShifts, DetectsNoShiftsIfEmpty) {
   // clang-format off
-  const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCTATAGATAACAGAACATTCTTGGAATGCTGATC" );
-  const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCMGTGGCTATAGATAANNNNNNNTTCTTGGAATGCTGATC" );
-  //                                                    M              NNNNNNN
+  const auto refAln = toNucleotideSequence( "" );
+  const auto qryAln = toNucleotideSequence( "" );
   // clang-format on
 
   const auto result = detectFrameShifts(refAln, qryAln);
@@ -34,32 +31,29 @@ TEST_F(DetectFrameShifts, DetectsNoShiftsIfNoIndels) {
   EXPECT_EQ(actual, expected);
 }
 
-TEST_F(DetectFrameShifts, DetectsNoShiftsIf3xDelsAdjacent) {
-  std::stringstream input;
+TEST_F(DetectFrameShifts, DetectsNoShiftsIfnoIndels) {
+  // clang-format off
+  const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCTATAGATAACAGAACATTCTTGGAATGCTGATC" );
+  const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCTATAGATAACAGAACATTCTTGGAATGCTGATC" );
+  // clang-format on
 
+  const auto result = detectFrameShifts(refAln, qryAln);
+  const auto& actual = result.frameShifts;
+
+  const std::vector<FrameShiftRange> expected = {};
+
+  EXPECT_EQ(actual, expected);
+}
+
+TEST_F(DetectFrameShifts, DetectsNoShiftsIf3xDelsetionsAdjacent) {
   // clang-format off
   //                                                   10        20        30        40        50
   //                                                   |         |         |         |         |
   //                                         012345678901234567890123456789012345678901234567890
   const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCTATAGATAACAGAACATTCTTGGAATGCTGATC" );
   const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCTAT---TAACAGAACATTCTTGGAATGCTGATC" );
-  // frame                                   00000000000000000000120000000000000000000000000000
-  // clang-format on
-
-  const auto result = detectFrameShifts(refAln, qryAln);
-  const auto& actual = result.frameShifts;
-
-  const std::vector<FrameShiftRange> expected = {};
-
-  EXPECT_EQ(actual, expected);
-}
-
-TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3) {
-  std::stringstream input;
-
-  // clang-format off
-  const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCTATAGATAACAGAACATTCTTGGAATGCTGATC" );
-  const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCCG---CTATAGATAACAGAACA---TTGGAATGCTGATC" );
+  // indel                                                       ddd
+  // frame                                                       120
   // clang-format on
 
   const auto result = detectFrameShifts(refAln, qryAln);
[…]
