### 2021-09-06 bf6ed44
feat: translate genes with length that is not divisible by 3
In case the length of a gene is not divisible by 3 we truncate it down to length modulo 3 and translate that.
