+++ packages/nextalign/src/translate/translateGenes.cpp
@@ -65,7 +65,6 @@ PeptidesInternal translateGenes(         //
     if (extractRefGeneStatus.status != Status::Success) {
       const auto message = *extractRefGeneStatus.error;
       warnings.inGenes.push_back(GeneWarning{.geneName = geneName, .message = message});
-      continue;
     }
 
 
@@ -77,14 +76,13 @@ PeptidesInternal translateGenes(         //
       if (extractQueryGeneStatus.reason == ExtractGeneStatusReason::GeneLengthNonMul3) {
         frameShifts.emplace_back(FrameShift{.geneName = geneName});
       }
-
-      continue;
     }
 
     const auto& refGene = *extractRefGeneStatus.result;
     const auto& queryGene = *extractQueryGeneStatus.result;
 
     const auto frameShiftResult = detectFrameShifts(refGene, queryGene);
+    const auto& frameShifts = frameShiftResult.frameShifts;
 
     auto refPeptide = translate(*extractRefGeneStatus.result);
     const auto queryPeptide = translate(*extractQueryGeneStatus.result);
