### 2021-09-06 9a3b19a
test: remove unused test fixture
