+++ packages/nextalign/tests/frameshiftDetector.test.cpp
@@ -8,16 +8,7 @@
 #include "../src/translate/detectFrameShifts.h"
 
 
-class DetectFrameShifts : public ::testing::Test {
-protected:
-  NextalignOptions options = getDefaultOptions();
-
-  DetectFrameShifts() {
-    options.alignment.minimalLength = 3;
-  }
-};
-
-TEST_F(DetectFrameShifts, DetectsNoShiftsIfEmpty) {
+TEST(DetectFrameShifts, DetectsNoShiftsIfEmpty) {
   // clang-format off
   const auto refAln = toNucleotideSequence( "" );
   const auto qryAln = toNucleotideSequence( "" );
@@ -31,7 +22,7 @@ TEST_F(DetectFrameShifts, DetectsNoShiftsIfEmpty) {
   EXPECT_EQ(actual, expected);
 }
 
-TEST_F(DetectFrameShifts, DetectsNoShiftsIfnoIndels) {
+TEST(DetectFrameShifts, DetectsNoShiftsIfnoIndels) {
   // clang-format off
   const auto refAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCTATAGATAACAGAACATTCTTGGAATGCTGATC" );
   const auto qryAln = toNucleotideSequence( "CTTGGAGGTTCCGTGGCTATAGATAACAGAACATTCTTGGAATGCTGATC" );
@@ -45,7 +36,7 @@ TEST_F(DetectFrameShifts, DetectsNoShiftsIfnoIndels) {
   EXPECT_EQ(actual, expected);
 }
 
-TEST_F(DetectFrameShifts, DetectsNoShiftsIf3xDelsetionsAdjacent) {
+TEST(DetectFrameShifts, DetectsNoShiftsIf3xDelsetionsAdjacent) {
   // clang-format off
   //                                                   10        20        30        40        50
   //                                                   |         |         |         |         |
@@ -64,7 +55,7 @@ TEST_F(DetectFrameShifts, DetectsNoShiftsIf3xDelsetionsAdjacent) {
   EXPECT_EQ(actual, expected);
 }
 
-TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3Adjacent) {
+TEST(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3Adjacent) {
   // clang-format off
   //                                                   10        20        30        40        50
   //                                                   |         |         |         |         |
@@ -83,7 +74,7 @@ TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3Adjacent) {
   EXPECT_EQ(actual, expected);
 }
 
-TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3Adjacent2) {
+TEST(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3Adjacent2) {
   // clang-format off
   //                                                   10        20        30        40        50
   //                                                   |         |         |         |         |
@@ -103,7 +94,7 @@ TEST_F(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3Adjacent2) {
   EXPECT_EQ(actual, expected);
 }
 
-TEST_F(DetectFrameShifts, DetectsNoShiftsIf3xInsertionsAdjacent) {
+TEST(DetectFrameShifts, DetectsNoShiftsIf3xInsertionsAdjacent) {
   // clang-format off
   //                                                   10        20        30        40        50
   //                                                   |         |         |         |         |
@@ -122,7 +113,7 @@ TEST_F(DetectFrameShifts, DetectsNoShiftsIf3xInsertionsAdjacent) {
   EXPECT_EQ(actual, expected);
 }
 
[…]
