### 2021-09-06 f803e22
feat: run frame shift detector on full aligned sequences
