### 2021-09-06 cbebafd
fix: detect frame shifts on unstripped genes
This moves gene stripping from the gene extraction place to after the frame shift detection is done. So the current flow is:
 - extract gene (unstriped)
 - detect frame shifts in gene
 - strip gene
 - translate
etc.
