+++ packages/nextalign/src/translate/extractGene.h
@@ -23,3 +23,5 @@ struct ExtractGeneStatus {
 
 ExtractGeneStatus extractGeneQuery(const NucleotideSequenceView& query, const Gene& gene,
   const std::vector<int>& coordMap);
+
+void stripGeneInPlace(NucleotideSequence& seq);
