### 2021-09-06 33721b2
feat: expose frame shifts in nextalign interface
They will be used in Nextclade CLi and Web, which use internal interface to Nextalign, but we don't plan to expose this to Nextalign CLI itself.
