### 2021-09-06 6d32233
fix: remove erroneous pre- and post-conditions in translation
It was still triggering in debug mode even though we translate non-modulo-3 genes since commit bf6ed44
