+++ packages/nextclade/src/io/serializeResults.cpp
@@ -104,6 +104,20 @@ namespace Nextclade {
       return j;
     }
 
+    json serializeFrameShiftRange(const FrameShiftRange& fs) {
+      auto j = json::object();
+      j.emplace("begin", fs.begin);
+      j.emplace("end", fs.end);
+      return j;
+    }
+
+    json serializeFrameShiftResult(const FrameShiftResult& fs) {
+      auto j = json::object();
+      j.emplace("geneName", fs.geneName);
+      j.emplace("frameShiftRange", serializeFrameShiftRange(fs.frameShiftRange));
+      return j;
+    }
+
     json serializeMissing(const NucleotideRange& missing) {
       return serializeNucleotideRange(missing);
     }
@@ -269,6 +283,7 @@ namespace Nextclade {
       j.emplace("totalSubstitutions", result.totalSubstitutions);
       j.emplace("totalDeletions", result.totalDeletions);
       j.emplace("totalInsertions", result.totalInsertions);
+      j.emplace("totalFrameShifts", result.totalFrameShifts);
       j.emplace("totalMissing", result.totalMissing);
       j.emplace("totalNonACGTNs", result.totalNonACGTNs);
       j.emplace("totalPcrPrimerChanges", result.totalPcrPrimerChanges);
@@ -279,6 +294,7 @@ namespace Nextclade {
       j.emplace("substitutions", serializeArray(result.substitutions, serializeMutation));
       j.emplace("deletions", serializeArray(result.deletions, serializeDeletion));
       j.emplace("insertions", serializeArray(result.insertions, serializeInsertion));
+      j.emplace("frameShifts", serializeArray(result.frameShifts, serializeFrameShiftResult));
       j.emplace("missing", serializeArray(result.missing, serializeMissing));
       j.emplace("nonACGTNs", serializeArray(result.nonACGTNs, serializeNonAcgtn));
       j.emplace("pcrPrimerChanges", serializeArray(result.pcrPrimerChanges, serializePcrPrimerChange));
@@ -299,6 +315,7 @@ namespace Nextclade {
     auto j = json::object();
     j.emplace("name", peptide.name);
     j.emplace("seq", peptide.seq);
+    j.emplace("frameShiftRanges", serializeArray(peptide.frameShiftRanges, serializeFrameShiftRange));
     return j;
   }
 
