+++ packages/nextclade/src/analyze/getFrameShifts.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <string>
+#include <vector>
+
+struct PeptideInternal;
+
+namespace Nextclade {
+  struct FrameShiftResult;
+
+  std::vector<FrameShiftResult> getFrameShifts(const std::vector<PeptideInternal>& queryPeptides);
+}// namespace Nextclade
