### 2021-09-07 bcc26fb
fix: protect only first codon before frame shift detection
So far we gap-stripped genes *after* frame shift detection. Right before stripping we would also protect the first and last not-all-gaps codons by replacing gap character `-` with `N` in them, in order to preserve reading frame during subsequent gap stripping. So the old flow was:
 - detect frame shifts
 - protect first and last codons
 - strip gaps

This was incorrect, because the frame shift detector have treated `-` characters in unprotected partial codons as deletions, which caused false-positive detections of frame shifts in the b
[…]
