+++ packages/nextalign/include/nextalign/private/nextalign_private.h
@@ -20,7 +20,7 @@ struct PeptideInternal {
   std::string name;
   AminoacidSequence seq;
   std::vector<InsertionInternal<Aminoacid>> insertions;
-  std::vector<FrameShiftRange> frameShiftRanges;
+  std::vector<FrameShiftResult> frameShiftResults;
 };
 
 
