### 2021-09-07 62f6bae
feat: make a coord map from alignment to reference (reverse coord map)
