### 2021-09-07 e346ade
fix: handle reverse map when there are leading insertions
