+++ packages/nextalign/tests/mapCoordinates.test.cpp
@@ -16,11 +16,21 @@ TEST(mapCoordinates, MapsSimple) {
 }
 
 TEST(mapCoordinates, MapsReverseSimple) {
-  // aln pos: 0  1  2  3  4  5  6  7  8  9  10 11 12 13 14
+  // ref pos: 0  1  2  3  4  5  6  7  8  9  10 11 12 13 14
   // ref    : A  C  T  C  -  -  -  C  G  T  G  -  -  -  A
-  // ref pos: 0  1  2  3  3  3  3  4  5  6  7  7  7  7  8
+  // aln pos: 0  1  2  3  3  3  3  4  5  6  7  7  7  7  8
 
   const auto ref = toNucleotideSequence("ACTC---CGTG---A");
   const auto expected = std::vector<int>{0, 1, 2, 3, 3, 3, 3, 4, 5, 6, 7, 7, 7, 7, 8};
+  EXPECT_THAT(mapReverseCoordinates(ref), expected);
+}
+
+TEST(mapCoordinates, MapsReverseSimpleWithLeadingInsertions) {
+  // ref pos: 0  1  2  3  4  5  6  7  8  9  10 11 12 13 14 15 16
+  // ref    : -  -  A  C  T  C  -  -  -  C  G  T  G  -  -  -  A
+  // aln pos: 0  0  0  1  2  3  3  3  3  4  5  6  7  7  7  7  8
+
+  const auto ref = toNucleotideSequence("--ACTC---CGTG---A");
+  const auto expected = std::vector<int>{0, 0, 0, 1, 2, 3, 3, 3, 3, 4, 5, 6, 7, 7, 7, 7, 8};
   EXPECT_THAT(mapReverseCoordinates(ref), expected);
 }
