### 2021-09-07 4b34fc4
test: adjust gene extraction tests to the new stripping logic
