### 2021-09-08 7c91301
refactor: remove unused code
