+++ packages/nextalign/tests/extractGene.test.cpp
@@ -9,32 +9,6 @@
 #include "../src/translate/mapCoordinates.h"
 
 
-TEST(extractGeneRef, ExtractsRefGene) {
-  // clang-format off
-  // gene                                    2          14
-  // gene                                    |           |
-  const std::string ref =                 "ACTCCGTGCGCTGCTG";
-  // const std::string ref_aligned =      "ACTCCG---TGCGCTGCTG";
-  // const std::string query_aligned =    "ACTGGTCTA---GCA";
-  const std::string expected_gene_ref =     "TCCGTGCGCTGC";
-  // const std::string expected_gene_query ="TGGTCTAGC";
-  // clang-format on
-
-  const Gene gene = {
-    .geneName = "Hello",
-    .start = 2,
-    .end = 14,
-    .strand = "+",
-    .frame = 0,
-    .length = 12,
-  };
-
-  const auto gene_ref = NucleotideSequence(extractGeneRef(toNucleotideSequence(ref), gene));
-
-  EXPECT_EQ(toString(gene_ref), expected_gene_ref);
-}
-
-
 TEST(extractGeneQuery, ExtractsQueryGene) {
   // clang-format off
   // gene                                    2          14
