### 2021-09-08 e31c083
fix: remove unnecessary length check
And skip the gene if there are still other problems
