### 2021-09-08 25a43cc
fix: properly use reverse map to convert frame shift ranges to ref coord
