+++ packages/nextalign/tests/translateGenes.test.cpp
@@ -1,0 +1,231 @@
+#include "../src/translate/translateGenes.h"
+
+#include <gtest/gtest.h>
+
+#include "../src/align/getGapOpenCloseScores.h"
+
+class TranslateGenes : public ::testing::Test {
+  NucleotideSequence ref;
+
+protected:
+  GeneMap geneMap;
+  NextalignOptions options = getDefaultOptions();
+  std::vector<int> gapOpenCloseAA;
+
+  TranslateGenes() {
+    options.alignment.minimalLength = 3;
+    // clang-format off
+    //                                      0         10        20        30        40        50        60        70        80        90        100       110       120
+    //                                      |         |         |         |         |         |         |         |         |         |         |         |         |
+    //                                      01234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901
+                ref = toNucleotideSequence("AGAAACTGCTCAAAATTCTGTGTGATATGAACAGAAGGCCGCTATAACAATACTACATGGAATTTCACTAGATTCACTGAGACTCATTGATGCTATGATGTTCACATTAGATTTGGCTAC");
+    //                                               |       Gene 1       |          |           Gene 2            |                 |      Gene 3        |
+    //                                               xxxxxxxxxxxxxxxxxxxxx           xxxxxxxxxxxxxxxxxxxxxxxxxxxxxx                  xxxxxxxxxxxxxxxxxxxxx
+    //                                                 9                    30       41                            71                99                   120
+    // clang-format on
+
+#pragma GCC diagnostic ignored "-Wmissing-field-initializers"
+    geneMap = GeneMap{
+      {"Gene1", Gene{.geneName = "Gene1", .start = 9, .end = 30, .frame = 0, .length = 21}},
+      {"Gene2", Gene{.geneName = "Gene2", .start = 41, .end = 71, .frame = 2, .length = 30}},
+      {"Gene3", Gene{.geneName = "Gene3", .start = 99, .end = 120, .frame = 0, .length = 21}},
+    };
+#pragma GCC diagnostic pop
+
+    gapOpenCloseAA = getGapOpenCloseScoresFlat(ref, options);
+  }
+};
+
+TEST_F(TranslateGenes, ExtractsAndTranslatesSimpleCase) {
+  // clang-format off
+  //                                        0         10        20        30        40        50        60        70        80        90        100       110       120
+  //                                        |         |         |         |         |         |         |         |         |         |         |         |         |
+  //                                        0123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890
+  //            ref = toNucleotideSequence("AGAAACTGCTCAAAATTCTGTGTGATATGAACAGAAGGCCGCTATAACAATACTACATGGAATTTCACTAGATTCACTGAGACTCATTGATGCTATGATGTTCACATTAGATTTGGCTAC");
+
+  //                                        0         10        20        30        40        50        60        70        80        90        100       110       120
+  //                                        |         |         |         |         |         |         |         |         |         |         |         |         |
+  //                                        0123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890
+  const auto refAln = toNucleotideSequence("AGAAACTGCTCAAAATTCTGTGTGATATGAACAGAAGGCCGCTATAACAATACTACATGGAATTTCACTAGATTCACTGAGACTCATTGATGCTATGATGTTCACATTAGATTTGGCTAC");
[…]
