+++ packages/nextalign/src/nextalign.cpp
@@ -45,7 +45,6 @@ NextalignResultInternal nextalignInternal(const NucleotideSequence& query, const
 
   std::vector<PeptideInternal> queryPeptides;
   std::vector<PeptideInternal> refPeptides;
-  std::vector<FrameShift> frameShifts;
   Warnings warnings;
   if (!geneMap.empty()) {
     auto peptidesInternal =
@@ -54,7 +53,6 @@ NextalignResultInternal nextalignInternal(const NucleotideSequence& query, const
     concat_move(peptidesInternal.refPeptides, refPeptides);
     concat_move(peptidesInternal.warnings.global, warnings.global);
     concat_move(peptidesInternal.warnings.inGenes, warnings.inGenes);
-    concat_move(peptidesInternal.frameShifts, frameShifts);
   }
 
   const auto stripped = stripInsertions(alignmentStatus.result->ref, alignmentStatus.result->query);
@@ -68,7 +66,6 @@ NextalignResultInternal nextalignInternal(const NucleotideSequence& query, const
     .queryPeptides = queryPeptides,
     .insertions = stripped.insertions,
     .warnings = warnings,
-    .frameShifts = frameShifts,
   };
 }
 
