+++ packages/nextalign/src/translate/translateGenes.cpp
@@ -44,7 +44,6 @@ PeptidesInternal translateGenes(         //
   refPeptides.reserve(geneMap.size());
 
   Warnings warnings;
-  std::vector<FrameShift> frameShifts;
 
   // For each gene in the requested subset
   for (const auto& [geneName, _] : geneMap) {
@@ -75,10 +74,6 @@ PeptidesInternal translateGenes(         //
     if (extractQueryGeneStatus.status != Status::Success) {
       const auto message = *extractQueryGeneStatus.error;
       warnings.inGenes.push_back(GeneWarning{.geneName = geneName, .message = message});
-
-      if (extractQueryGeneStatus.reason == ExtractGeneStatusReason::GeneLengthNonMul3) {
-        frameShifts.emplace_back(FrameShift{.geneName = geneName});
-      }
       continue;
     }
 
@@ -135,6 +130,5 @@ PeptidesInternal translateGenes(         //
     .queryPeptides = queryPeptides,//
     .refPeptides = refPeptides,    //
     .warnings = warnings,          //
-    .frameShifts = frameShifts,    //
   };
 }
