+++ packages/nextalign/src/translate/translateGenes.h
@@ -10,7 +10,6 @@ struct PeptidesInternal {
   std::vector<PeptideInternal> queryPeptides;
   std::vector<PeptideInternal> refPeptides;
   Warnings warnings;
-  std::vector<FrameShift> frameShifts;
 };
 
 PeptidesInternal translateGenes(         //
