### 2021-09-08 8de881c
chore: uncomment temporarily commented tests
