### 2021-09-09 bace7e5
refactor: deduplicate Range and FrameShiftRange types
