+++ packages/nextalign/tests/frameshiftDetector.test.cpp
@@ -16,7 +16,7 @@ TEST(DetectFrameShifts, DetectsNoShiftsIfEmpty) {
 
   const auto actual = detectFrameShifts(refAln, qryAln);
 
-  const std::vector<FrameShiftRange> expected = {};
+  const std::vector<Range> expected = {};
 
   EXPECT_EQ(actual, expected);
 }
@@ -29,7 +29,7 @@ TEST(DetectFrameShifts, DetectsNoShiftsIfnoIndels) {
 
   const auto actual = detectFrameShifts(refAln, qryAln);
 
-  const std::vector<FrameShiftRange> expected = {};
+  const std::vector<Range> expected = {};
 
   EXPECT_EQ(actual, expected);
 }
@@ -47,7 +47,7 @@ TEST(DetectFrameShifts, DetectsNoShiftsIf3xDelsetionsAdjacent) {
 
   const auto actual = detectFrameShifts(refAln, qryAln);
 
-  const std::vector<FrameShiftRange> expected = {};
+  const std::vector<Range> expected = {};
 
   EXPECT_EQ(actual, expected);
 }
@@ -65,7 +65,7 @@ TEST(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3Adjacent) {
 
   const auto actual = detectFrameShifts(refAln, qryAln);
 
-  const std::vector<FrameShiftRange> expected = {};
+  const std::vector<Range> expected = {};
 
   EXPECT_EQ(actual, expected);
 }
@@ -84,7 +84,7 @@ TEST(DetectFrameShifts, DetectsNoShiftsIfDeletionsAreModulo3Adjacent2) {
 
   const auto actual = detectFrameShifts(refAln, qryAln);
 
-  const std::vector<FrameShiftRange> expected = {};
+  const std::vector<Range> expected = {};
 
   EXPECT_EQ(actual, expected);
 }
@@ -102,7 +102,7 @@ TEST(DetectFrameShifts, DetectsNoShiftsIf3xInsertionsAdjacent) {
 
   const auto actual = detectFrameShifts(refAln, qryAln);
 
-  const std::vector<FrameShiftRange> expected = {};
+  const std::vector<Range> expected = {};
 
   EXPECT_EQ(actual, expected);
 }
@@ -120,7 +120,7 @@ TEST(DetectFrameShifts, DetectsNoShiftsIfInsertionsAreModulo3Adjacent) {
 
   const auto actual = detectFrameShifts(refAln, qryAln);
 
-  const std::vector<FrameShiftRange> expected = {};
+  const std::vector<Range> expected = {};
 
   EXPECT_EQ(actual, expected);
 }
@@ -138,7 +138,7 @@ TEST(DetectFrameShifts, DetectsNoShiftsIfInsertionsAreModulo3Adjacent2) {
 
   const auto actual = detectFrameShifts(refAln, qryAln);
 
-  const std::vector<FrameShiftRange> expected = {};
+  const std::vector<Range> expected = {};
 
   EXPECT_EQ(actual, expected);
 }
@@ -156,7 +156,7 @@ TEST(DetectFrameShifts, DetectsNoShiftsIfDeletionsAndInsertionsAreModulo3AndAdja
 
   const auto actual = detectFrameShifts(refAln, qryAln);
 
-  const std::vector<FrameShiftRange> expected = {};
+  const std::vector<Range> expected = {};
 
   EXPECT_EQ(actual, expected);
 }
@@ -174,7 +174,7 @@ TEST(DetectFrameShifts, DetectsNoShiftsIfInsertionsAndDeletionsAreModulo3AndAdja
 
   const auto actual = detectFrameShifts(refAln, qryAln);
 
-  const std::vector<FrameShiftRange> expected = {};
[…]

[…]
