+++ packages/nextclade/CMakeLists.txt
@@ -109,7 +109,6 @@ set(SOURCE_FILES
   src/utils/map.h
   src/utils/mapFind.h
   src/utils/range.h
-  src/utils/removeInPlace.h
   src/utils/safe_cast.h
   )
 
