### 2021-09-09 4afd5e2
fix: correct aa changes filtering logic
