### 2021-09-09 f80f424
feat: mask frame shift ranges in genes
This replaces non-gap characters in frame shift ranges in nuc sequence of each gene with N. This way the shift consists entirely from aminoacids X. This helps peptide alignment to not produce huge insertions.

The masking goes before stripping gaps, to be able to operate on relative nuc positions (relative to gene start and in alignment crdinates). This is why we keep gaps unmasked.
