### 2021-09-09 0997fee
feat: mask frame shift ranges in peptides
Followup of the  f80f424fd6d3f7bec2668c7a9c580a2e4741f5d5

This masks gaps in frame shifted regions of the peptide with aminoacid X.
These regions are misaligned, so these gaps added during peptide alignment don't make sense.
