### 2021-09-09 a299a5e
feat(web): calm down nucleotide colors
