### 2021-09-14 c1244ac
feat: reimplement dataset selection flow
