### 2021-09-16 #3691 @kevinlibuit
Kelly,

Yes! Would be more than happy to contribute to the review of this
manuscript.

Best,

Kevin

On Mon, Sep 13, 2021 at 10:29 AM Kelly L. Rowland ***@***.***>
wrote:

> Thanks for the reviewer suggestions, @rneher <https://github.com/rneher> !
>
> Hello @pvanheus <https://github.com/pvanheus> @kevinlibuit
> <https://github.com/kevinlibuit> @EduanWilkinson
> <https://github.com/EduanWilkinson> 👋 - would you be interested in and
> available to review this paper for JOSS? We require a minimum of two
> reviewers per paper.
>
