### 2021-09-20 759c50b
feat: ignore predefined frame shifts in qc
