+++ packages/nextclade/src/io/parseAnalysisResults.cpp
@@ -240,6 +240,13 @@ namespace Nextclade {
     };
   }
 
+  FrameShiftLocation parseFrameShiftLocation(const json& j) {
+    return FrameShiftLocation{
+      .geneName = at(j, "geneName"),
+      .codonRange = parseRange(at(j, "codonRange")),
+    };
+  }
+
   std::optional<QcResultFrameShifts> parseQcFrameShifts(const json& jj) {
     if (!jj.contains("frameShifts")) {
       return {};
@@ -248,8 +255,10 @@ namespace Nextclade {
     return QcResultFrameShifts{
       .score = parseDouble(j, "score"),
       .status = parseQcStatus(frozen::string{j["status"].get<std::string>()}),
-      .frameShifts = parseArray<FrameShiftResult>(j, "frameShifts", parseFrameShiftResult),
-      .totalFrameShifts = parseInt(j, "totalFrameShifts"),
+      .frameShiftsReported = parseArray<FrameShiftResult>(j, "frameShiftsReported", parseFrameShiftResult),
+      .totalFrameShiftsReported = parseInt(j, "totalFrameShiftsReported"),
+      .frameShiftsIgnored = parseArray<FrameShiftResult>(j, "frameShiftsIgnored", parseFrameShiftResult),
+      .totalFrameShiftsIgnored = parseInt(j, "totalFrameShiftsIgnored"),
     };
   }
 
