### 2021-09-20 d045d8b
refactor: rename field to avoid breaking changes
Let's keep `frameShifts` for reported frame shifts, instead of newly introduced `frameShiftsReported`. Ignored frame shifts stay in `totalFrameShiftsIgnored`.
