### 2021-09-20 757a8be
refactor: rename field to avoid breaking changes
Let's keep `frameShifts` for reported frame shifts, instead of newly introduced `frameShiftsReported`. Ignored frame shifts stay in `totalFrameShiftsIgnored`.

This is a second attempt after the incorrect attempt in d045d8ba69094e20c6143123de78749ee31b1da3
