### 2021-09-20 11fa720
chore: reenable nextalign run
