### 2021-09-20 192e4b3
feat: add `--no-translate-past-stop` flag to Nextalign CLI
