### 2021-09-21 a81a9f2
feat(algo): report ignored stop codons separately
