### 2021-09-21 b3d7922
fix: adjust unit tests for the frame shift codon masking
The masking was introduced earlier in 8095d6192ce4d4ad4dac4f9ea4289ecbda26e55b and bc40677d6869f4e9542f7b130a4deaac6f1d5ce8 

The tests are adjusted to expect X amino acids within the frame shift body instead of the actual translation.
