### 2021-09-23 5e43a8b
fix: mask codons containing nuc deletions before and after frame shift
