### 2021-09-24 d94b531
fix: find begin of the preceding deletion correctly
