+++ packages/nextalign/src/translate/detectFrameShifts.h
@@ -4,13 +4,14 @@
 
 
 std::vector<Range> detectFrameShifts(//
-  const NucleotideSequence& ref,               //
-  const NucleotideSequence& query              //
+  const NucleotideSequence& ref,     //
+  const NucleotideSequence& query    //
 );
 
-std::vector<FrameShiftResult> translateFrameShifts(     //
-  const std::vector<Range>& nucRelFrameShifts,//
-  const std::vector<int>& coordMap,                     //
-  const std::vector<int>& coordMapReverse,              //
-  const Gene& gene                                      //
+std::vector<InternalFrameShiftResultWithMask> translateFrameShifts(//
+  const NucleotideSequence& query,                                 //
+  const std::vector<Range>& nucRelFrameShifts,                     //
+  const std::vector<int>& coordMap,                                //
+  const std::vector<int>& coordMapReverse,                         //
+  const Gene& gene                                                 //
 );
