+++ packages/nextalign/include/nextalign/private/nextalign_private.h
@@ -93,13 +93,24 @@ inline std::ostream& operator<<(std::ostream& os, const Range& f) {
   return os;
 }
 
+inline std::ostream& operator<<(std::ostream& os, const FrameShiftContext& f) {
+  os << "{ "                        //
+     << "codon: " << f.codon << ", "//
+     << " }"                        //
+    ;
+  return os;
+}
+
+
 inline std::ostream& operator<<(std::ostream& os, const FrameShiftResult& f) {
-  os << "{ "                                  //
-     << "geneName: \"" << f.geneName << "\", "//
-     << "nucRel: " << f.nucRel << ", "        //
-     << "nucAbs: " << f.nucAbs << ", "        //
-     << "codon: " << f.codon << ", "          //
-     << " }"                                  //
+  os << "{ "                                    //
+     << "geneName: \"" << f.geneName << "\", "  //
+     << "nucRel: " << f.nucRel << ", "          //
+     << "nucAbs: " << f.nucAbs << ", "          //
+     << "codon: " << f.codon << ", "            //
+     << "gapsLeading: " << f.gapsLeading << ", "//
+     << "gapsTraling: " << f.gapsTraling << ", "//
+     << " }"                                    //
     ;
   return os;
 }
