### 2021-09-24 8f104bf
fix: fox off-by-1 error in trailing deletion detection, add tests
