### 2021-09-24 f6bdf91
fix: add leading and trailing gaps functionality to tests
