### 2021-09-24 5075827
feat: emit leading and trailing gaps into JSON
