### 2021-09-24 fffd990
feat(web): display leading and trailing gaps in frame shift tooltips
