### 2021-09-24 9d2be49
feat: mask leading/trailing aa deletions with gap character
