+++ packages/nextalign/src/translate/detectFrameShifts.cpp
@@ -146,7 +146,7 @@ int findMaskBegin(const NucleotideSequence& seq, const Range& frameShiftNucRange
     }
   }
 
-  // `begin` now points to the nuc that is 1 to the right of the deletion.
+  // `begin` now points to the nuc that is 1 to the left of the deletion.
   // Go back one nuc to make it point to the deletion.
   return begin + 1;
 }
@@ -166,7 +166,7 @@ int findMaskEnd(const NucleotideSequence& seq, const Range& frameShiftNucRangeRe
     }
   }
 
-  // `end` now points to the nuc that is 1 to the left of the deletion. Which is correct - we use semi-open ranges.
+  // `end` now points to the nuc that is 1 to the right of the deletion. Which is correct - we use semi-open ranges.
   return end;
 }
 
