### 2021-09-24 487349c
fix: ensure correct invariants for index when accessing an array
