### 2021-09-27 91a9c96
fix: add missing lookup to the coord map
