### 2021-09-27 cf09d15
refactor: simplify loops
