### 2021-09-27 #3691 @kellyrowland
Hi @dcnickle @Maghnuso - would either of you be willing and available to review this submission to JOSS? I see that you were the reviewers on the paper for the Augur software package, which is used and referenced here.
