### 2021-09-27 #3691 @DavidNickle
Yes - that was an error. Let’s stick with ***@***.*** ***@***.***>.

Thank you,

David

> On Sep 27, 2021, at 1:37 PM, Kelly L. Rowland ***@***.***> wrote:
>
>
> @DavidNickle <https://github.com/DavidNickle> - I just caught that your account here in this issue is different than the one I originally pinged. Do you have a preference for which to use as the reviewer assignment?
>
