### 2021-09-27 #3691 @kellyrowland
Yep, I can see the email address, thanks. The JOSS workflow bot expects Github handles directly as input, however - is either account okay to reach you at that address? I don't mean to belabor the question, just trying to make sure everything's lined up correctly here before we move things over to the review issue.
