### 2021-09-27 #3691 @DavidNickle
Yes either is fine.  I am currently working for a small biotech that uses Git Lab and I think I confused the system by logging in on the wrong account.  But both work.

Thanks,

David

> On Sep 27, 2021, at 2:19 PM, Kelly L. Rowland ***@***.***> wrote:
>
>
> Yep, I can see the email address, thanks. The JOSS workflow bot expects Github handles directly as input, however - is either account okay to reach you at that address? I don't mean to belabor the question, just trying to make sure everything's lined up correctly here before we move things over to the review issue.
>
