### 2021-09-28 681e60d
refactor: encapsulate conversion of seq positions in a class
