+++ packages/nextalign/src/translate/mapCoordinates.cpp
@@ -8,56 +8,151 @@
 #include "utils/contract.h"
 #include "utils/safe_cast.h"
 
-/**
- * Makes a map from reference coordinates to alignment coordinates
- * (Excluding the gaps in reference).
- *
- * Example:
- *   ref pos: 0  1  2  3  4  5  6  7  8  9  10 11 12 13 14
- *   ref    : A  C  T  C  -  -  -  C  G  T  G  -  -  -  A
- *   aln pos: 0  1  2  3           7  8  9  10          14
- */
-std::vector<int> mapCoordinates(const NucleotideSequence& ref) {
-  const auto refLength = safe_cast<int>(ref.size());
 
-  std::vector<int> coordMap;
-  coordMap.reserve(refLength);
-  for (int i = 0; i < refLength; ++i) {
-    if (ref[i] != Nucleotide::GAP) {
-      coordMap.push_back(i);
+namespace {
+
+  /*****************************************************************************
+   *
+   *                 OLD IMPLEMENTATIONS
+   *
+   ****************************************************************************/
+
+
+  /**
+ * Makes the "alingnment to reference" coordinate map: from alignment coordinates to reference coordinates.
+ *
+ * Given a position of a letter in the aligned sequence, the "alingnment to reference" coordinate map allows to
+ * lookup the position of the corresponding letter in the reference sequence.
+ *
+ * @param refAln  Aligned reference sequence
+ * @return        Coordinate map from alignment coordinates to reference coordinates
+ */
+  std::vector<int> makeAlnToRefMapOld(const NucleotideSequence& ref) {
+    const auto alnLength = safe_cast<int>(ref.size());
+
+    std::vector<int> revCoordMap;
+    revCoordMap.reserve(alnLength);
+    int refPos = 0;
+    for (int i = 0; i < alnLength; ++i) {
+      if (ref[i] == Nucleotide::GAP) {
+        const auto& prev = revCoordMap.back();
+        revCoordMap.push_back(prev);
+      } else {
+        revCoordMap.push_back(refPos);
+        ++refPos;
+      }
     }
+
+    revCoordMap.shrink_to_fit();
+    return revCoordMap;
   }
 
-  coordMap.shrink_to_fit();
-  return coordMap;
+
+  /**
+ * Makes the "reference to alingnment" coordinate map: from alignment coordinates to reference coordinates.
+ *
+ * Given a position of a letter in the reference sequence, the "reference to alingnment" coordinate map allows to
+ * lookup the position of the corresponding letter in the aligned sequence.
+ *
+ * @param refAln  Aligned reference sequence
+ * @return        Coordinate map from alignment coordinates to reference coordinates
+ */
+  std::vector<int> makeRefToAlnMapOld(const NucleotideSequence& ref) {
+    const auto alnLength = safe_cast<int>(ref.size());
+
+    std::vector<int> coordMap;
+    coordMap.reserve(alnLength);
+    for (int i = 0; i < alnLength; ++i) {
+      if (ref[i] != Nucleotide::GAP) {
+        coordMap.push_back(i);
+      }
+    }
+
+    coordMap.shrink_to_fit();
+    return coordMap;
+  }
+
+
+  /*****************************************************************************
+   *
+   *                 NEW IMPLEMENTATIONS
+   *
+   ****************************************************************************/
+
+
+  /**
+ * Makes the "alingnment to reference" coordinate map: from alignment coordinates to reference coordinates.
+ *
[…]
