### 2021-09-28 #3691 @Maghnuso
Hi Kelly,
Are you still looking for another reviewer? I’m on parental leave at the moment so I might pass on this one, but if you are in a bind, do let me know and I can probably fit it in.
Best,
Maghnus

Sent from my iPhone

> On Sep 27, 2021, at 16:24, Kelly L. Rowland ***@***.***> wrote:
>
> ﻿
> Hi @dcnickle @Maghnuso - would either of you be willing and available to review this submission to JOSS? I see that you were the reviewers on the paper for the Augur software package, which is used and referenced here.
>
