### 2021-09-28 2fddf72
chore: add JOSS badge to README
