+++ README.md
@@ -135,6 +135,16 @@ by Nextstrain team
 
 </p>
 
+<p align="center">
+
+  <a href="https://joss.theoj.org/papers/b0e8e7f231ddfc61616de6b56f32b79f">
+    <img 
+      src="https://joss.theoj.org/papers/b0e8e7f231ddfc61616de6b56f32b79f/status.svg">
+      alt="JOSS submission status"
+  </a>
+
+</p>
+
 ---
 
 <p align="center">
