### 2021-09-28 #3773 @DavidNickle
**Conflict of interest:** I confirm that I have read the JOSS conflict of interest policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

Code of Conduct

**Repository:** The source code for this software is available at the repository url
clicking through the link Nextclade CLI
https://docs.nextstrain.org/projects/nextclade/en/stable/user/nextclade-cli.html

**License:** The repository contains a plain-text LICENSE file with the contents of an OSI approved software license.
All source code is freely available under the terms of the GNU Affero General Public License.

**Contribution and authorship:**  It is not clear who did what for the paper

**Functionality:** The software is functional and important adding very important knowledge to the public domain - I particular like the alignment algorithm takes information about different readings.

**Installation:** The installation procedure follows the documentation clearly - I followed the Nextclade CLI method and ran it the SARS-CoV-2 example. From:
`./nextclade dataset get --name='sars-cov-2' --output-dir='data/sars-cov-2'`

**Functionality:** I confirmed the functional claims of the on the tutorial data as well as some of my own.

**Performance:** It appears to quick! But the authors did not claim anything around speed.

**A statement of need:** The authors have a clear audience in mine and the paper and the software should be readably available to that audience.

**Installation instructions:** Following the command line instructions for Nextclade CLI works as claimed. I fallowed the Mac instructions.

**Example usage:** I went through the the SARS-CoV-2 example and all the functionality worked. I found exercise easy and straight forward.
`$ ./nextclade run --input-dataset=data/sars-cov-2/ --input-fasta=data/sars-cov-2/sequences.fasta`

**Automated tests:** YES — tests are available
Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. YES - but they should start with Nextstrain base install. This is declared in the Nextclade CLI install page.

**Software paper:** YES

**Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided? YES

**A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is? YES

**State of the field:** Do the authors describe how this software compares to other commonly-used packages? YES
Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? YES
References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax? YES

Lastly, the banded pairwise Smith-Waterman alignment is a nice idea and I can imaging implementing that method in mofft or another multiple sequence aligner.

This should be accepted as is.
