### 2021-09-28 ff4b2b2
feat: add arithmetic operations on ranges
