+++ packages/nextalign/include/nextalign/nextalign.h
@@ -35,13 +35,33 @@ struct Range {
   int begin;
   int end;
 
-  bool contains(int x) const {
+  [[nodiscard]] bool contains(int x) const {
     return x >= begin && x < end;
   }
 };
 
-inline bool operator==(const Range& left, const Range& right) {
+[[nodiscard]] inline bool operator==(const Range& left, const Range& right) {
   return left.begin == right.begin && left.end == right.end;
+}
+
+[[nodiscard]] inline bool operator!=(const Range& left, const Range& right) {
+  return left.begin != right.begin && left.end != right.end;
+}
+
+[[nodiscard]] inline Range operator+(const Range& r, int add) {
+  return Range{.begin = r.begin + add, .end = r.end + add};
+}
+
+[[nodiscard]] inline Range operator-(const Range& r, int sub) {
+  return Range{.begin = r.begin - sub, .end = r.end - sub};
+}
+
+[[nodiscard]] inline Range operator*(const Range& r, int mul) {
+  return Range{.begin = r.begin * mul, .end = r.end * mul};
+}
+
+[[nodiscard]] inline Range operator/(const Range& r, int div) {
+  return Range{.begin = r.begin / div, .end = r.end / div};
 }
 
 struct FrameShiftContext {
@@ -63,11 +83,11 @@ struct FrameShiftResult {
 };
 
 inline bool operator==(const FrameShiftResult& left, const FrameShiftResult& right) {
-  return left.geneName == right.geneName         //
-         && left.nucRel == right.nucRel          //
-         && left.nucAbs == right.nucAbs          //
-         && left.codon == right.codon            //
-         && left.gapsLeading == right.gapsLeading//
+  return left.geneName == right.geneName           //
+         && left.nucRel == right.nucRel            //
+         && left.nucAbs == right.nucAbs            //
+         && left.codon == right.codon              //
+         && left.gapsLeading == right.gapsLeading  //
          && left.gapsTrailing == right.gapsTrailing//
     ;
 }
