+++ packages/nextalign/src/translate/mapCoordinates.cpp
@@ -156,3 +156,17 @@ int CoordinateMapper::alnToRef(int alnPos) const {
 int CoordinateMapper::refToAln(int refPos) const {
   return at(refToAlnMap, refPos);
 }
+
+Range CoordinateMapper::alnToRef(const Range& alnRange) const {
+  return Range{
+    .begin = alnToRef(alnRange.begin),
+    .end = alnToRef(alnRange.end),
+  };
+}
+
+Range CoordinateMapper::refToAln(const Range& refRange) const {
+  return Range{
+    .begin = refToAln(refRange.begin),
+    .end = refToAln(refRange.end),
+  };
+}
