### 2021-09-28 0b3c3c6
refactor: simplify coordinate conversion for frame shifts and masks
