### 2021-09-28 4ef6690
fix: mask peptides before stripping insertions from it
