### 2021-09-28 5e8c50b
Revert "fix: mask peptides before stripping insertions from it"
This reverts commit 4ef66906bc6fe15ff1c524c6a50f9e1dacd1e2ab.
