### 2021-09-29 ce73f8f
fix: account for indels outside of genes when converting ranges
