### 2021-09-29 5c1277b
fix: convert positions in peptides to account for peptide insertions
