+++ packages/nextalign/src/translate/extractGene.h
@@ -6,10 +6,10 @@
 #include <string>
 #include <vector>
 
+#include "mapCoordinates.h"
 #include "nextalign/private/nextalign_private.h"
 
 struct Gene;
-class CoordinateMapper;
 
 enum ExtractGeneStatusReason { GeneEmpty };
 
@@ -21,6 +21,6 @@ struct ExtractGeneStatus {
 };
 
 ExtractGeneStatus extractGeneQuery(const NucleotideSequenceView& query, const Gene& gene,
-  const CoordinateMapper& coordMap);
+  const CoordinateMapper<Nucleotide>& coordMap);
 
 void protectFirstCodonInPlace(NucleotideSequence& seq);
