+++ packages/nextalign/src/translate/mapCoordinates.cpp
@@ -7,166 +7,3 @@
 
 #include "utils/contract.h"
 #include "utils/safe_cast.h"
-
-
-namespace {
-
-  /*****************************************************************************
-   *
-   *                 OLD IMPLEMENTATIONS
-   *
-   ****************************************************************************/
-
-
-  /**
- * Makes the "alingnment to reference" coordinate map: from alignment coordinates to reference coordinates.
- *
- * Given a position of a letter in the aligned sequence, the "alingnment to reference" coordinate map allows to
- * lookup the position of the corresponding letter in the reference sequence.
- *
- * @param refAln  Aligned reference sequence
- * @return        Coordinate map from alignment coordinates to reference coordinates
- */
-  std::vector<int> makeAlnToRefMapOld(const NucleotideSequence& ref) {
-    const auto alnLength = safe_cast<int>(ref.size());
-
-    std::vector<int> revCoordMap;
-    revCoordMap.reserve(alnLength);
-    int refPos = 0;
-    for (int i = 0; i < alnLength; ++i) {
-      if (ref[i] == Nucleotide::GAP) {
-        const auto& prev = revCoordMap.back();
-        revCoordMap.push_back(prev);
-      } else {
-        revCoordMap.push_back(refPos);
-        ++refPos;
-      }
-    }
-
-    revCoordMap.shrink_to_fit();
-    return revCoordMap;
-  }
-
-
-  /**
- * Makes the "reference to alingnment" coordinate map: from alignment coordinates to reference coordinates.
- *
- * Given a position of a letter in the reference sequence, the "reference to alingnment" coordinate map allows to
- * lookup the position of the corresponding letter in the aligned sequence.
- *
- * @param refAln  Aligned reference sequence
- * @return        Coordinate map from alignment coordinates to reference coordinates
- */
-  std::vector<int> makeRefToAlnMapOld(const NucleotideSequence& ref) {
-    const auto alnLength = safe_cast<int>(ref.size());
-
-    std::vector<int> coordMap;
-    coordMap.reserve(alnLength);
-    for (int i = 0; i < alnLength; ++i) {
-      if (ref[i] != Nucleotide::GAP) {
-        coordMap.push_back(i);
-      }
-    }
-
-    coordMap.shrink_to_fit();
-    return coordMap;
-  }
-
-
-  /*****************************************************************************
-   *
-   *                 NEW IMPLEMENTATIONS
-   *
-   ****************************************************************************/
-
-
-  /**
- * Makes the "alingnment to reference" coordinate map: from alignment coordinates to reference coordinates.
- *
- * Given a position of a letter in the aligned sequence, the "alingnment to reference" coordinate map allows to
- * lookup the position of the corresponding letter in the reference sequence.
- *
- * @param refAln  Aligned reference sequence
- * @return        Coordinate map from alignment coordinates to reference coordinates
- */
-  std::vector<int> makeAlnToRefMap(const NucleotideSequence& refAln) {
-    int alnLength = safe_cast<int>(refAln.size());
-    std::vector<int> alnToRefMap(alnLength);// TODO: Slow
-    int refPos = -1;
-    for (int alnPos = 0; alnPos < alnLength; ++alnPos) {
-      if (refAln[alnPos] != Nucleotide::GAP) {
-        refPos += 1;
-      }
-      alnToRefMap[alnPos] = refPos;
-    }
-    return alnToRefMap;
-  }
-
-  /**
- * Makes the "reference to alingnment" coordinate map: from alignment coordinates to reference coordinates.
- *
[…]

[…]
