### 2021-09-29 7de23e7
fix: incorrect invariants
