+++ packages/nextalign/src/utils/at.h
@@ -10,3 +10,11 @@ inline const typename Container::value_type& at(const Container& container, Inde
   invariant_less(i, container.size());
   return container[i];
 }
+
+template<typename Container, typename Index>
+inline typename Container::value_type& at(Container& container, Index index) {
+  const auto i = safe_cast<typename Container::size_type>(index);
+  invariant_greater_equal(i, 0);
+  invariant_less(i, container.size());
+  return container[i];
+}
