### 2021-09-29 c19dbdd
fix: pass gene seq, not fill seq to calculation of mask
