### 2021-09-29 0bc52a0
test: adjust tests for the masking
The masking with gaps was introduced in https://github.com/nextstrain/nextclade/pull/549
and broke a few expectations on peptide output in tests. This commit adjusts these expectations.
