### 2021-09-29 #3691 @kellyrowland
Hi @EduanWilkinson we can add you as a third reviewer over in the review issue, sure.

The workflow would go as follows:

In the review issue at https://github.com/openjournals/joss-reviews/issues/3773 , I would @ whedon, the JOSS bot, and have it add you as a reviewer on the paper. This will send you an invite to the repository, which you'll need to accept to proceed with the review process.

Once you've been added as a reviewer and accepted the repo invite, you'll be assigned to the Github issue, where you'll have a set of 20 checkboxes to work through. You can take a look at the checkboxes in the issue linked previously; each reviewer is asked to work through this same set of checkboxes to ensure software functionality and documentation and review the associated paper.

Any questions or feedback for the authors can be raised on the software repo or in the review issue; reviewers and authors are encouraged to iterate as needed to arrive at a software and paper state appropriate for publication in JOSS. Once all reviewers have completed their reviews, I'll have the authors release a tagged version of the software to go with the paper and then ping the JOSS editors-in-chief to finish out the publication.

Please let me know if you have any questions about the above and if you are interested in and able to carry out the review - we can add you as a third reviewer on the submission.
