### 2021-09-30 c0e2a27
fix(CHANGELOG): typos
