### 2021-09-30 db45c55
chore(Changelog): edit and add information to frame shift detection part
