### 2021-09-30 25607eb
fix: typo in comments in alignPairwise.cpp
