### 2021-09-30 6f163df
docs(changelog): clarify ui changes and behavior, remove judgement
