### 2021-09-30 38a0b41
docs(changelog): rewrite to clarify the changes and add more background
