### 2021-09-30 de81881
docs(changelog): remove duplicated section
