### 2021-10-05 530c463
fix: reimplement csv writer
Resolves: https://github.com/nextstrain/nextclade/issues/560

This is a complete reimplementation of CsvWriter class in a more robust way.

Previously this class used simple, in-house implementation. This change transitions the class to use `rapidcsv` library.

Depennencies:
 - 3rdparty package`rapidcsv` was added
 - 3rdparty package`frozen` had to be updated to incorporate this fix in `frozen::map` https://github.com/serge-sans-paille/frozen/issues/116
