+++ 3rdparty/frozen/include/frozen/map.h
@@ -27,6 +27,7 @@
 #include "frozen/bits/basic_types.h"
 #include "frozen/bits/constexpr_assert.h"
 #include "frozen/bits/exceptions.h"
+#include "frozen/bits/type_traits.h"
 #include "frozen/bits/version.h"
 
 #include <utility>
@@ -48,10 +49,22 @@ public:
                            std::pair<Key, Value> const &other) const {
     return comparator_(std::get<0>(self), std::get<0>(other));
   }
+  template <class Key1, class Key2, class Value,
+            typename = bits::has_is_transparent_t<Comparator, Key1>>
+  constexpr int operator()(std::pair<Key1, Value> const &self,
+                           std::pair<Key2, Value> const &other) const {
+    return comparator_(std::get<0>(self), std::get<0>(other));
+  }
 
   template <class Key, class Value>
   constexpr int operator()(Key const &self_key,
                            std::pair<Key, Value> const &other) const {
+    return comparator_(self_key, std::get<0>(other));
+  }
+  template <class Key1, class Key2, class Value,
+            typename = bits::has_is_transparent_t<Comparator, Key1>>
+  constexpr int operator()(Key1 const &self_key,
+                           std::pair<Key2, Value> const &other) const {
     return comparator_(self_key, std::get<0>(other));
   }
 
@@ -60,9 +73,20 @@ public:
                            Key const &other_key) const {
     return comparator_(std::get<0>(self), other_key);
   }
+  template <class Key1, class Key2, class Value,
+            typename = bits::has_is_transparent_t<Comparator, Key1>>
+  constexpr int operator()(std::pair<Key1, Value> const &self,
+                           Key2 const &other_key) const {
+    return comparator_(std::get<0>(self), other_key);
+  }
 
   template <class Key>
   constexpr int operator()(Key const &self_key, Key const &other_key) const {
+    return comparator_(self_key, other_key);
+  }
+  template <class Key1, class Key2,
+            typename = bits::has_is_transparent_t<Comparator, Key1>>
+  constexpr int operator()(Key1 const &self_key, Key2 const &other_key) const {
     return comparator_(self_key, other_key);
   }
 };
@@ -110,7 +134,7 @@ public:
       : map{items, Compare{}} {}
 
   /* element access */
-  constexpr Value& at(Key const &key) const {
+  constexpr Value const& at(Key const &key) const {
     return at_impl(*this, key);
   }
   constexpr Value& at(Key const &key) {
@@ -142,12 +166,27 @@ public:
   constexpr std::size_t count(Key const &key) const {
     return bits::binary_search<N>(items_.begin(), key, less_than_);
   }
+  template <class K,
+            typename = bits::has_is_transparent_t<Compare, K>>
+  constexpr std::size_t count(K const &key) const {
+    return bits::binary_search<N>(items_.begin(), key, less_than_);
+  }
 
   constexpr const_iterator find(Key const &key) const {
     return find_impl(*this, key);
   }
   constexpr iterator find(Key const &key) {
     return find_impl(*this, key);
+  }
+  template <class K,
+            typename = bits::has_is_transparent_t<Compare, K>>
+  constexpr const_iterator find(K const &key) const {
+    return map::find_impl(*this, key);
+  }
+  template <class K,
+            typename = bits::has_is_transparent_t<Compare, K>>
+  constexpr iterator find(K const &key) {
+    return map::find_impl(*this, key);
   }
 
   constexpr std::pair<const_iterator, const_iterator>
@@ -157,11 +196,32 @@ public:
   constexpr std::pair<iterator, iterator> equal_range(Key const &key) {
     return equal_range_impl(*this, key);
   }
[…]
