+++ 3rdparty/rapidcsv/tests/test008.cpp
@@ -1,0 +1,49 @@
+// test008.cpp - delete rows
+
+#include <rapidcsv.h>
+#include "unittest.h"
+
+int main()
+{
+  int rv = 0;
+
+  std::string csvref =
+    "-,A,B,C,D\n"
+    "1,3,9,81,6561\n"
+    "2,4,16,256,65536\n"
+  ;
+
+  std::string csv =
+    "-,A,B,C,D\n"
+    "0,2,4,16,256\n"
+    "1,3,9,81,6561\n"
+    "2,4,16,256,65536\n"
+    "3,5,25,625,390625\n"
+  ;
+
+  std::string path = unittest::TempPath();
+  unittest::WriteFile(path, csv);
+
+  try
+  {
+    rapidcsv::Document doc(path, rapidcsv::LabelParams(0, 0));
+
+    doc.RemoveRow("3");
+    doc.RemoveRow(0);
+
+    doc.Save();
+
+    std::string csvread = unittest::ReadFile(path);
+
+    unittest::ExpectEqual(std::string, csvref, csvread);
+  }
+  catch (const std::exception& ex)
+  {
+    std::cout << ex.what() << std::endl;
+    rv = 1;
+  }
+
+  unittest::DeleteFile(path);
+
+  return rv;
+}
