### 2021-10-05 4822be6
fix: adjust wasm build for the new CSVWriter class
