### 2021-10-06 311d3de
chore: avoid running executables and linting csv when cross-compiling
