### 2021-10-06 21ac3c9
chore: add plausible to csp
