+++ packages/web/infra/lambda-at-edge/modifyOutgoingHeaders.lambda.js
@@ -1,4 +1,3 @@
-/* eslint-disable prefer-destructuring */
 // Adds additional headers to the response, including security headers.
 // Suited for websites.
 //
@@ -31,7 +30,7 @@ function generateFeaturePolicyHeader(featurePoicyObject) {
 
 const NEW_HEADERS = {
   'Content-Security-Policy':
-    "default-src 'self'; script-src 'self' 'unsafe-eval'; style-src 'self' 'unsafe-inline'; img-src 'self' data: *.githubusercontent.com; connect-src *",
+    "default-src 'self'; script-src 'self' 'unsafe-eval' plausible.io; style-src 'self' 'unsafe-inline'; img-src 'self' data: *.githubusercontent.com; connect-src *",
   'Referrer-Policy': 'no-referrer',
   'Strict-Transport-Security': 'max-age=15768000; includeSubDomains; preload',
   'X-Content-Type-Options': 'nosniff',
