### 2021-10-07 ec76721
fix: add missing serialization for fields in analysis results
This code is used to marshal data from C++ to JS and lack of serialization for these fileds was causing missing aa mutations on the tree
