### 2021-10-08 50d1355
refactor: hoist ref sequence and ref peptide processing out of main loop
