### 2021-10-11 29c034c
chore: add snapshot tests for cli
