### 2021-10-12 #3773 @kevinlibuit
@davekinkead I think that we are supposed to provide feedback by modifying the checklist above. [Here is an example of a JOSS review completed by a colleague of mine](https://github.com/openjournals/joss-reviews/issues/1390).

Though I'm still unable to update the checklist. @kellyrowland is it alright if I proceed with my review in a similar way as David, i.e. provide text response to each item in the checklist?
