+++ packages/nextclade/src/__tests__/getDifferences.test.cpp
@@ -1,0 +1,171 @@
+#include <fmt/format.h>
+#include <gmock/gmock.h>
+#include <gtest/gtest.h>
+
+#include <exception>
+#include <fstream>
+
+#include "../../include/nextclade/nextclade.h"
+#include "../../include/nextclade/private/nextclade_private.h"
+#include "../../src/io/parseMutation.h"
+#include "../../src/tree/TreeNode.h"
+#include "../../src/tree/treeAttachNodes.h"
+
+
+#define EXPECT_ARR_EQ_UNORDERED(expected, actual) ASSERT_THAT(actual, ::testing::UnorderedElementsAreArray(expected))
+
+namespace {
+  using Nextclade::AnalysisResult;
+  using Nextclade::NucleotideSubstitution;
+  using Nextclade::Tree;
+  using Nextclade::TreeNode;
+
+  using Nextclade::getDifferences;
+  using Nextclade::parseMutation;
+
+
+  class getDifferences : public ::testing::Test {
+  protected:
+    NucleotideSequence rootSeq;
+
+    NucleotideSequence parseRefFastaFile(const std::string& filename) {
+      std::ifstream file(filename);
+      if (!file.good()) {
+        throw std::runtime_error(fmt::format("Error: unable to read \"{:s}\"\n", filename));
+      }
+
+      const auto refSeqs = parseSequences(file, filename);
+      if (refSeqs.size() != 1) {
+        throw std::runtime_error(
+          fmt::format("Error: {:d} sequences found in reference sequence file, expected 1", refSeqs.size()));
+      }
+
+      const auto& refSeq = refSeqs.front();
+      return toNucleotideSequence(refSeq.seq);
+    }
+
+    getDifferences() {
+      rootSeq = parseRefFastaFile("data/sars-cov-2/reference.fasta");
+    }
+
+    TreeNode makeRef(Tree& tree, const std::vector<std::string>& substitutions) {
+
+
+      auto node = tree.root();
+
+      std::map<int, Nucleotide> substitutionsMap;
+      for (const auto& sub : substitutions) {
+        const auto mut = parseMutation(sub);
+        substitutionsMap.insert(std::make_pair(mut.pos, mut.queryNuc));
+      }
+      node.setSubstitutions(substitutionsMap);
+
+      return node;
+    }
+
+    std::vector<NucleotideSubstitution> makeMutList(const std::vector<std::string>& substitutions) {
+      std::vector<NucleotideSubstitution> muts;
+      for (const auto& sub : substitutions) {
+        const auto mut = parseMutation(sub);
+        muts.emplace_back(mut);
+      }
+      return muts;
+    }
+
+    AnalysisResult makeQuery(const std::vector<std::string>& substitutions) {
+      return AnalysisResult{.substitutions = makeMutList(substitutions)};
+    }
+
+    std::vector<NucleotideSubstitution> toStrings(
+      const std::map<std::string, std::vector<std::string>>& substitutions) {
+
+      std::vector<NucleotideSubstitution> muts;
+      for (const auto& sub : substitutions.at("nuc")) {
+        const auto mut = parseMutation(sub);
+        muts.emplace_back(mut);
+      }
+      return muts;
+
+      //      return substitutions.at("nuc");
+      //      std::vector<std::string> muts;
+      //      for (const auto& sub : substitutions.at("nuc")) {
+      //        const auto mut = Nextclade::formatMutation(sub);
[…]
