### 2021-10-18 05663ff
refactor: move max divergence and div units calculation to preprocessing
Here we move max divergence and divergence units calculation from "tree attach nodes" to "tree preprocessing" step. This information then can be used by the "tree find nearest node" step. This is needed in order to perform private mutation detection and new node divergence calculation during "tree find nearest node" step, which can be done concurrently. Currently these are performed in a sequential  "tree attach nodes" step, so it makes Nextclade somewhat slower. This order of operations also makes code somewhat clearer, 
[…]
