### 2021-10-18 55da078
refactor: deduplicate private mutations calculation code
