+++ packages/nextclade/src/tree/calculateDivergence.h
@@ -1,0 +1,14 @@
+#pragma once
+
+namespace Nextclade {
+  struct AnalysisResult;
+  class TreeNode;
+  enum class DivergenceUnits;
+
+  double calculateDivergence(       //
+    const TreeNode& nearestNode,    //
+    const AnalysisResult& result,   //
+    DivergenceUnits divergenceUnits,//
+    int refSeqLength                //
+  );
+}//namespace Nextclade
