+++ packages/nextclade/src/tree/treePreprocess.h
@@ -4,6 +4,13 @@
 
 namespace Nextclade {
   class Tree;
+  class TreeNode;
+
+  std::map<int, Nucleotide> mapNucleotideMutations(      //
+    const TreeNode& node,                                //
+    const NucleotideSequence& rootSeq,                   //
+    const std::map<int, Nucleotide>& parentNucMutationMap//
+  );
 
   void treePreprocess(Tree& tree, const NucleotideSequence& rootSeq,
     const std::map<std::string, RefPeptideInternal>& refPeptides);
