+++ packages/nextclade/src/tree/treeAttachNodes.h
@@ -7,14 +7,5 @@
 namespace Nextclade {
   class Tree;
 
-  void treeAttachNodes(Tree& tree, const NucleotideSequence& rootSeq, const std::vector<AnalysisResult>& results);
-
-  struct GetDifferencesResult {
-    std::map<std::string, std::vector<std::string>> mutations;
-    std::vector<std::string> nucMutations;
-    double divergence;
-  };
-
-  GetDifferencesResult getDifferences(const AnalysisResult& result, const TreeNode& node,
-    const NucleotideSequence& rootSeq, double maxDivergence);
+  void treeAttachNodes(Tree& tree, const std::vector<AnalysisResult>& results);
 }// namespace Nextclade
