+++ packages/nextclade/src/qc/ruleMixedSites.h
@@ -5,13 +5,11 @@
 
 namespace Nextclade {
   struct AnalysisResult;
-  struct NucleotideSubstitution;
   struct QCResultMixedSites;
   struct QCRulesConfigMixedSites;
 
-  std::optional<QCResultMixedSites> ruleMixedSites(      //
-    const AnalysisResult& result,                       //
-    const std::vector<NucleotideSubstitution>& mutations,//
-    const QCRulesConfigMixedSites& config                //
+  std::optional<QCResultMixedSites> ruleMixedSites(//
+    const AnalysisResult& result,                  //
+    const QCRulesConfigMixedSites& config          //
   );
 }// namespace Nextclade

[…]
