+++ packages/nextclade/src/tree/findPrivateNucMutations.cpp
@@ -10,23 +10,46 @@
 #include "utils/mapFind.h"
 
 namespace Nextclade {
+  /**
+   * Finds private nucleotide mutations.
+   *
+   * Private mutations are the mutations in the query (user-provided) sequence relative to the parent node on the
+   * reference tree. At this point sequence have not yet become new node on the tree, but is just described by the
+   * results of the analysis.
+   *
+   * We have an array of sequence mutations relative to reference and a map of node mutations relative to reference.
+   * We want to find private mutations. The following cases are possible:
+   * |---|--------------------------------------------------------------|---------|---------------|--------------|
+   * |   |                          Case                                | Private |      From     |     To       |
+   * |---|--------------------------------------------------------------|---------|---------------|--------------|
+   * | 1 | mutation in sequence and in node, same query character       |   no    |               |              |
+   * |---|--------------------------------------------------------------|---------|---------------|--------------|
+   * | 2 | mutation in sequence and in node, but not the same character |   yes   | node.queryNuc | seq.queryNuc |
+   * |---|--------------------------------------------------------------|---------|---------------|--------------|
+   * | 3 | mutation in sequence but not in node                         |   yes   | seq.refNuc    | seq.queryNuc |
+   * |---|--------------------------------------------------------------|---------|---------------|--------------|
+   * | 4 | mutation in node, but not in sequence                        |   yes   | node.queryNuc | node.refNuc  |
+   * |   | (mutation in sequence that reverts the character to ref)     |         |               |              |
+   * |---|--------------------------------------------------------------|---------|---------------|--------------|
+   */
   PrivateMutations findPrivateNucMutations(     //
     const std::map<int, Nucleotide>& nodeMutMap,//
-    const AnalysisResult& result,               //
+    const AnalysisResult& seq,                  //
     const NucleotideSequence& refSeq            //
   ) {
+
     std::vector<NucleotideSubstitutionSimple> privateSubstitutions;
-    privateSubstitutions.reserve(result.totalSubstitutions);
+    privateSubstitutions.reserve(seq.substitutions.size());
 
     std::vector<NucleotideDeletionSimple> privateDeletions;
-    privateDeletions.reserve(result.totalDeletions);
+    privateDeletions.reserve(seq.deletions.size());
 
-    std::set<int> positionsCovered;
+    std::set<int> seqPositionsCovered;
 
-    for (const auto& qmut : result.substitutions) {
+    for (const auto& qmut : seq.substitutions) {
       const auto& pos = qmut.pos;
       const auto& queryNuc = qmut.queryNuc;
-      positionsCovered.insert(pos);
+      seqPositionsCovered.insert(pos);
       const auto der = mapFind(nodeMutMap, pos);
 
       std::optional<Nucleotide> refNuc;
@@ -49,19 +72,19 @@ namespace Nextclade {
       }
     }
 
-    for (const auto& del : result.deletions) {
+    for (const auto& del : seq.deletions) {
       const auto& start = del.start;
       const auto& end = del.start + del.length;
 
       for (int pos = start; pos < end; ++pos) {
-        const auto& der = mapFind(nodeMutMap, pos);
-        positionsCovered.insert(pos);
+        const auto& nodeQueryNuc = mapFind(nodeMutMap, pos);
+        seqPositionsCovered.insert(pos);
 
         std::optional<Nucleotide> refNuc;
-        if (der) {
-          if (isGap(*der)) {
+        if (nodeQueryNuc) {
+          if (isGap(*nodeQueryNuc)) {
[…]
