### 2021-10-19 3a2dabf
refactor: simplify and test findPrivateNucMutations function
