+++ packages/nextclade/src/tree/findPrivateNucMutations.cpp
@@ -5,9 +5,10 @@
 #include <map>
 #include <vector>
 
-#include "analyze/isSequenced.h"
-#include "analyze/nucleotide.h"
-#include "utils/mapFind.h"
+#include "../analyze/isSequenced.h"
+#include "../analyze/nucleotide.h"
+#include "../utils/eraseDuplicates.h"
+#include "../utils/mapFind.h"
 
 namespace Nextclade {
   /**
@@ -22,14 +23,14 @@ namespace Nextclade {
    * |---|--------------------------------------------------------------|---------|---------------|--------------|
    * |   |                          Case                                | Private |      From     |     To       |
    * |---|--------------------------------------------------------------|---------|---------------|--------------|
-   * | 1 | mutation in sequence and in node, same query character       |   no    |               |              |
+   * | 1 | mutation in sequence and in node, same query character       |   no    |      N/A      |     N/A      |
    * |---|--------------------------------------------------------------|---------|---------------|--------------|
    * | 2 | mutation in sequence and in node, but not the same character |   yes   | node.queryNuc | seq.queryNuc |
    * |---|--------------------------------------------------------------|---------|---------------|--------------|
    * | 3 | mutation in sequence but not in node                         |   yes   | seq.refNuc    | seq.queryNuc |
    * |---|--------------------------------------------------------------|---------|---------------|--------------|
    * | 4 | mutation in node, but not in sequence                        |   yes   | node.queryNuc | node.refNuc  |
-   * |   | (mutation in sequence that reverts the character to ref)     |         |               |              |
+   * |   | (mutation in sequence that reverts the character to ref seq) |         |               |              |
    * |---|--------------------------------------------------------------|---------|---------------|--------------|
    */
   PrivateMutations findPrivateNucMutations(     //
@@ -39,39 +40,38 @@ namespace Nextclade {
   ) {
 
     std::vector<NucleotideSubstitutionSimple> privateSubstitutions;
-    privateSubstitutions.reserve(seq.substitutions.size());
+    privateSubstitutions.reserve(seq.totalSubstitutions + nodeMutMap.size());
 
     std::vector<NucleotideDeletionSimple> privateDeletions;
-    privateDeletions.reserve(seq.deletions.size());
+    privateDeletions.reserve(seq.totalDeletions + nodeMutMap.size());
 
+    // Remember which positions we cover while iterating sequence mutations,
+    // to be able to skip them when we iterate over node mutations
     std::set<int> seqPositionsCovered;
 
-    for (const auto& qmut : seq.substitutions) {
-      const auto& pos = qmut.pos;
-      const auto& queryNuc = qmut.queryNuc;
+    // Process sequence substitutions
+    for (const auto& seqMut : seq.substitutions) {
+      const auto& pos = seqMut.pos;
+      const auto& nodeQueryNuc = mapFind(nodeMutMap, pos);
       seqPositionsCovered.insert(pos);
-      const auto der = mapFind(nodeMutMap, pos);
 
-      std::optional<Nucleotide> refNuc;
-      if (der) {
-        if (queryNuc != der) {
-          // shared site but states of node and seq differ
-          refNuc = der;
-        }
-      } else {
-        // node does not have a mutation, but seq does -> compare to root
-        refNuc = refSeq[pos];
+      if (!nodeQueryNuc) {
+        // Case 3: Mutation in sequence but not in node, i.e. a newly occurred mutation.
[…]
