### 2021-10-20 f915a82
feat: add proper search for private aminoacid mutations
The bulk of the diff is refactoring: In order to share code for mutation detection needed to make the aminoacid and nucleotide structs generic and compatible, such that they could be passed to template specializations.
