+++ packages/nextclade/src/analyze/aminoacid.cpp
@@ -1,9 +1,0 @@
-#include "aminoacid.h"
-
-
-namespace Nextclade {
-  bool isGap(Aminoacid nuc) {
-    return nuc == Aminoacid::GAP;
-  }
-
-}// namespace Nextclade
