+++ packages/nextclade/src/tree/findPrivateNucMutations.cpp
@@ -1,5 +1,6 @@
 #include "findPrivateNucMutations.h"
 
+#include <fmt/format.h>
 #include <nextclade/nextclade.h>
 
 #include <map>
@@ -9,8 +10,156 @@
 #include "../analyze/nucleotide.h"
 #include "../utils/eraseDuplicates.h"
 #include "../utils/mapFind.h"
+#include "utils/concat_move.h"
 
 namespace Nextclade {
+  template<typename Letter>
+  struct LetterTag {};
+
+  template<typename Letter>
+  bool isSequencedGeneric(int pos, const AnalysisResult& seq, LetterTag<Letter>);
+
+  template<>
+  bool isSequencedGeneric(int pos, const AnalysisResult& seq, LetterTag<Nucleotide>) {
+    return isSequenced(pos, seq);
+  }
+
+  template<>
+  bool isSequencedGeneric(int /*pos*/, const AnalysisResult& /*seq*/, LetterTag<Aminoacid>) {
+    // For aminoacid sequences there is no concept of being not sequences currently. They are always sequenced.
+    return true;
+  }
+
+  template<typename Letter>
+  void processSeqSubstitutions(                                               //
+    /* in */ const std::map<int, Letter>& nodeMutMap,                         //
+    /* in */ const std::vector<Substitution<Letter>>& substitutions,          //
+    /* inout */ std::vector<SubstitutionSimple<Letter>>& privateSubstitutions,//
+    /* inout */ std::set<int>& seqPositionsCovered                            //
+  ) {
+    for (const auto& seqMut : substitutions) {
+      const auto& pos = seqMut.pos;
+      const auto& nodeQueryChar = mapFind(nodeMutMap, pos);
+      seqPositionsCovered.insert(pos);
+
+      if (!nodeQueryChar) {
+        // Case 3: Mutation in sequence but not in node, i.e. a newly occurred mutation.
+        // Action: Add the sequence mutation itself.
+        privateSubstitutions.emplace_back(SubstitutionSimple<Letter>{.ref = seqMut.ref, .pos = pos, .qry = seqMut.qry});
+      } else if (seqMut.qry != nodeQueryChar) {
+        // Case 2: Mutation in sequence and in node, but the query character is not the same.
+        // Action: Add mutation from node query character to sequence query character.
+        privateSubstitutions.emplace_back(
+          SubstitutionSimple<Letter>{.ref = *nodeQueryChar, .pos = pos, .qry = seqMut.qry});
+      }
+
+      // Otherwise case 1: mutation in sequence and in node, same query character, i.e. the mutation is not private:
+      // nothing to do.
+    }
+  }
+
+  /**
+   * Find private deletions corresponding to the deletions that sequence has.
+   * This is a general template. It has 2 specializations below: for nucleotide sequences and for aminoacid sequences.
+   * This is due to deletions having different data structure for nucleotides and for amino acids (range vs point).
+   */
+  template<typename Letter>
+  void processSeqDeletions(                                           //
+    /* in */ const std::map<int, Letter>& nodeMutMap,                 //
+    /* in */ const std::vector<Deletion<Letter>>& deletions,          //
+    /* in */ const Sequence<Letter>& refSeq,                          //
+    /* inout */ std::vector<DeletionSimple<Letter>>& privateDeletions,//
+    /* inout */ std::set<int>& seqPositionsCovered                    //
+  );
+
+  template<>
+  void processSeqDeletions(                                               //
+    /* in */ const std::map<int, Nucleotide>& nodeMutMap,                 //
+    /* in */ const std::vector<Deletion<Nucleotide>>& deletions,          //
+    /* in */ const Sequence<Nucleotide>& refSeq,                          //
[…]
