### 2021-10-20 99d13c5
feat: add deletions in node to general test case
