### 2021-10-20 d5f85df
chore: aligned boxes in docstring properly
