### 2021-10-20 c5a4015
feat: propagate changes in private aminoacid mut detection to web
