+++ packages/nextclade_json/include/nextclade_json/nextclade_json.h
@@ -63,6 +63,20 @@ namespace Nextclade {
     return parseArray<T>(arr, parser);
   }
 
+  template<typename Key, typename Val, typename Parser>
+  std::map<Key, Val> parseMap(const json& parent, const std::string& parentKey, const Parser& parser) {
+    const auto& obj = at(parent, parentKey);
+    if (!obj.is_object()) {
+      throw ErrorJsonTypeInvalid(parentKey, "object", parent.type_name());
+    }
+
+    std::map<Key, Val> result;
+    for (const auto& [key, value] : obj.items()) {
+      result[key] = parser(value);
+    }
+    return result;
+  }
+
 
   std::string jsonStringify(const json& j, int spaces = 2);
 }// namespace Nextclade
