### 2021-10-25 3408465
fix: fix crash in dataset list command on macos
