### 2021-10-25 f91247b
chore: [temp] produce macos executable on ci
