### 2021-10-25 8586bed
fix: install the latest conan through pip on macos
https://github.com/conan-io/conan-center-index/issues/5197
