+++ .circleci/config.yml
@@ -43,7 +43,7 @@ jobs:
       - restore_cache:
           name: 'Restore previous checksum.txt from cache'
           keys:
-            - cache-nextclade-v6-builder-container-{{ checksum "checksum.current.txt" }}
+            - cache-nextclade-v7-builder-container-{{ checksum "checksum.current.txt" }}
 
       - run:
           name: 'Compare checksum.txt to cached one, halt if no changes'
@@ -80,7 +80,7 @@ jobs:
 
       - save_cache:
           name: 'Save new checksum.txt to cache'
-          key: cache-nextclade-v6-builder-container-{{ checksum "checksum.previous.txt" }}
+          key: cache-nextclade-v7-builder-container-{{ checksum "checksum.previous.txt" }}
           paths:
             - checksum.previous.txt
 
@@ -101,7 +101,7 @@ jobs:
       - restore_cache:
           name: 'Restore WebAssembly module build cache'
           keys:
-            - cache-nextclade-v6-web-wasm-{{ arch }}-{{ checksum "conanfile.wasm.txt" }}
+            - cache-nextclade-v7-web-wasm-{{ arch }}-{{ checksum "conanfile.wasm.txt" }}
 
       - run:
           name: 'Build WebAssembly module'
@@ -112,14 +112,14 @@ jobs:
 
       - save_cache:
           name: 'Save WebAssembly module build cache'
-          key: cache-nextclade-v6-web-wasm-{{ arch }}-{{ checksum "conanfile.wasm.txt" }}
+          key: cache-nextclade-v7-web-wasm-{{ arch }}-{{ checksum "conanfile.wasm.txt" }}
           paths:
             - .cache
 
       - restore_cache:
           name: 'Restore web application build cache'
           keys:
-            - cache-nextclade-v6-web-js-{{ arch }}-{{ checksum "packages/web/yarn.lock" }}
+            - cache-nextclade-v7-web-js-{{ arch }}-{{ checksum "packages/web/yarn.lock" }}
 
       - run:
           name: 'Build web application'
@@ -128,7 +128,7 @@ jobs:
 
       - save_cache:
           name: 'Save web application build cache'
-          key: cache-nextclade-v6-web-js-{{ arch }}-{{ checksum "packages/web/yarn.lock" }}
+          key: cache-nextclade-v7-web-js-{{ arch }}-{{ checksum "packages/web/yarn.lock" }}
           paths:
             - packages/web/.cache
             - packages/web/.build/production/tmp/cache
@@ -188,7 +188,7 @@ jobs:
       - restore_cache:
           name: 'Restore build cache'
           keys:
-            - cache-nextclade-v6-cache-{{ arch }}-{{ checksum "conanfile.txt" }}
+            - cache-nextclade-v7-cache-{{ arch }}-{{ checksum "conanfile.txt" }}
 
       - run:
           name: 'Run "Builder" container'
@@ -198,7 +198,7 @@ jobs:
 
       - save_cache:
           name: 'Save build cache'
-          key: cache-nextclade-v6-cache-{{ arch }}-{{ checksum "conanfile.txt" }}
+          key: cache-nextclade-v7-cache-{{ arch }}-{{ checksum "conanfile.txt" }}
           paths:
             - .cache
 
@@ -242,7 +242,7 @@ jobs:
       - restore_cache:
           name: 'Restore homebrew cache'
           keys:
-            - cache-nextclade-v6-homebrew-{{ arch }}-{{ checksum "Brewfile" }}
+            - cache-nextclade-v7-homebrew-{{ arch }}-{{ checksum "Brewfile" }}
 
       - run:
[…]
