### 2021-10-25 fce4f8d
refactor: upgrade fmt
This version introduces bugfixes and related api changes
