### 2021-10-25 2fcb478
chore: only consider fully static build on linux
