### 2021-10-25 143a25a
chore: only strip executables on linux
Stripping on mac produced unexpected crashes. This needs to be investigated, but meanwhile, let's remove the stripping.
