### 2021-10-26 2611e2b
docs: remove `=` from CLI examples to align with built-in `--help`
