### 2021-10-27 5b8d687
fix: make sure aa mutations map is properly created
