### 2021-10-27 65908d0
docs(paper): setup build for preprint pdf
