### 2021-10-27 931d2b6
docs(paper): reduce margin in the preprint
