### 2021-10-27 7edcfee
docs(paper): setup build for preprint pdf - joss version
