+++ Makefile
@@ -208,46 +208,27 @@ docker-paper:
 
 paper-preprint:
 	@set -euxo pipefail
-
 	@cd paper/
-
-	pandoc \
-		--verbose \
-		--filter pandoc-citeproc \
-		--bibliography=paper.bib \
-		--variable classoption=twocolumn \
-		--variable papersize=a4paper \
-		--variable geometry:margin=2cm \
-		--standalone \
-		--output preprint.pdf \
-		paper.md
+	./scripts/build_preprint.sh
 
 docker-paper-preprint:
 	@set -euxo pipefail
 
-	@export CONTAINER_IMAGE_NAME=nextclade-paper-builder
+	@export CONTAINER_IMAGE_NAME=nextclade-preprint-builder
 
 	@docker build -t "$${CONTAINER_IMAGE_NAME}" \
 	--network=host \
-	--build-arg USER=$(shell id -un) \
-	--build-arg GROUP=$(shell id -gn) \
-	--build-arg UID=$(shell id -u) \
-	--build-arg GID=$(shell id -g) \
 	paper/
 
 	@docker run -it --rm \
 	--init \
 	--name="$${CONTAINER_IMAGE_NAME}-$(shell date +%s)" \
-	--user="$(shell id -un):$(shell id -gn)" \
+	--user="$(shell id -u):$(shell id -g)" \
 	--volume="$(shell pwd):/home/user/src" \
 	--workdir="/home/user/src" \
 	--env "TERM=xterm-256colors" \
-	--env "USER=$(shell id -un)" \
-	--env "GROUP=$(shell id -gn)" \
-	--env "UID=$(shell id -u)" \
-	--env "GID=$(shell id -g)" \
 	"$${CONTAINER_IMAGE_NAME}" \
-		make paper-preprint
+		bash -c "make paper-preprint"
 
 # Synchronize source files using rsync
 sync:
