### 2021-10-27 0e2063a
docs(paper): remove logo from preprint
