### 2021-10-27 f4e1522
docs(paper): expose latex template
