### 2021-10-27 491f0b3
docs(paper): add pandoc lua plugins for the preprint
