### 2021-10-27 0744868
docs(paper): expose latex template for the preprint
