### 2021-10-27 nextstrain/nextclade#584 @corneliusroemer
[docs] Nextalign quick example uses outdated way to download example dataset
Spotted by @kevinlibuit here: https://github.com/openjournals/joss-reviews/issues/3773#issuecomment-953199551

> [I] noticed that the NextAlign quick example suggests grabbing data files from the defunct data/sars-cov-2 directory.
>Would be great if there was a similar nextalign dataset get functionality to grab reference data. Not something super critical to my recommendation for acceptance to JOSS, but this would be very useful for laboratories running these from the nextstrain/nextalign docker image--especially in some kind of larger Augur workflow.

See point 1 in the screenshot below, it should use `nextclade dataset get` or `nextalign dataset get` but I don't think nextalign CLI has this feature.

![image](https://user-images.githubusercontent.com/25161793/139142846-4c3de18c-23c5-4361-82b2-b7fbe035c591.png)

https://docs.nextstrain.org/projects/nextclade/en/stable/user/nextalign-cli.html#quick-example
