### 2021-10-27 bdad2c2
feat: point docs to non-deprecated dataset repository
Temporary fix, we may possibly want to implement a better way to get example datasets for nextalign
