### 2021-10-27 9505f5f
chore: update auspice version automatically
