+++ .github/workflows/Update-Auspice.yml
@@ -1,0 +1,101 @@
+name: 'Update Auspice'
+
+on:
+  schedule:
+    - cron: '0 8 * * *'
+
+  repository_dispatch:
+    types: Update-Auspice
+
+  workflow_dispatch:
+
+
+jobs:
+  Update-Auspice:
+    runs-on: ubuntu-20.04
+
+    env:
+      GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 1
+
+      - uses: actions/setup-node@v1
+        with:
+          node-version: 14
+
+      - name: "Check Auspice version"
+        id: check-auspice-version
+        shell: bash
+        run: |
+          set -euxo pipefail
+          PATH="$HOME/.local/bin:$PWD/node_modules/.bin:$PATH"
+
+          npm install semver
+
+          # Get latest version of Auspice on NPM
+          export AUSPICE_VERSION="$(npm show auspice version)"
+
+          # Check if latest version of Auspice on NPM is greater than what's in Nextclade's package.json
+          export SHOULD_UPDATE=$(node -e "\
+            const semver = require('semver'); \
+            const pkg = require('./packages/web/package.json'); \
+            console.log(semver.gt(${AUSPICE_VERSION}, pkg.dependencies.auspice)) \
+          ")
+
+          echo "::set-output name=AUSPICE_VERSION::${AUSPICE_VERSION}"
+          echo "::set-output name=SHOULD_UPDATE::${SHOULD_UPDATE}"
+
+      - name: "Update Auspice version"
+        id: update-auspice-version
+        shell: bash
+        if: steps.check-auspice-version.outputs.SHOULD_UPDATE == 'true'
+        run: |
+          set -euxo pipefail
+          PATH="$HOME/.local/bin:$PWD/node_modules/.bin:$PATH"
+
+          git config --global user.name nextstrain-bot
+          git config --global user.email 78992647+nextstrain-bot@users.noreply.github.com
+
+          cd packages/web/
+
+          yarn install auspice@${AUSPICE_VERSION}
+
+          git add package.json yarn.lock
+          git commit -m "feat: update Auspice to version ${{ steps.check-auspice-version.outputs.AUSPICE_VERSION }}"
+
+      - name: "Create Pull Request"
+        id: create-pull-request
+        if: steps.check-auspice-version.outputs.SHOULD_UPDATE == 'true'
+        uses: peter-evans/create-pull-request@v3
+        with:
+          token: ${{ secrets.NEXTSTRAIN_BOT_GITHUB_TOKEN }}
+          author: 'nextstrain-bot <nextstrain-bot@users.noreply.github.com>'
+          committer: 'nextstrain-bot <nextstrain-bot@users.noreply.github.com>'
+          branch: 'chore/update-auspice'
+          commit-message: "feat: update Auspice"
+          title: '[bot] Update Auspice"'
+          body: |
+            Updates Auspice to version ${{ steps.check-auspice-version.outputs.AUSPICE_VERSION }}.
+
+            The changelog for the new version should be here: https://github.com/nextstrain/auspice/releases/tag/v${{ steps.check-auspice-version.outputs.AUSPICE_VERSION }}
+
+            Please check the `/tree` page of the Nextclade Web app. An automated Vercel deployment should be available below. Note that when even newer version is available, this PR will be overwritten, so a new check will be needed.
+
+            This automated Pull Request is generated by GitHub Action [${{github.run_id}}](https://github.com/nextstrain/nextclade/actions/runs/${{github.run_id}}?check_suite_focus=true).
[…]
