### 2021-10-27 5327932
chore: fix unbound var in GitHub Action
