+++ .github/workflows/Update-Auspice.yml
@@ -61,10 +61,12 @@ jobs:
 
           cd packages/web/
 
+          export AUSPICE_VERSION="${{ steps.check-auspice-version.outputs.AUSPICE_VERSION }}"
+
           yarn install auspice@${AUSPICE_VERSION}
 
           git add package.json yarn.lock
-          git commit -m "feat: update Auspice to version ${{ steps.check-auspice-version.outputs.AUSPICE_VERSION }}"
+          git commit -m "feat: update Auspice to version ${AUSPICE_VERSION}"
 
       - name: "Create Pull Request"
         id: create-pull-request
@@ -96,6 +98,8 @@ jobs:
           delete-branch: true
 
       - name: Check outputs
+        id: check-outputs
+        if: steps.check-auspice-version.outputs.SHOULD_UPDATE == 'true'
         run: |
           echo "Pull Request Number - ${{ steps.create-pull-request.outputs.pull-request-number }}"
           echo "Pull Request URL - ${{ steps.create-pull-request.outputs.pull-request-url }}"
