### 2021-10-27 6aa3bc0
chore: fix yarn command var in GitHub Action
