### 2021-10-27 e33c9d1
chore: cleanup unrelated files in GitHub Action
