+++ .github/workflows/Update-Auspice.yml
@@ -57,16 +57,22 @@ jobs:
           PATH="$HOME/.local/bin:$PWD/node_modules/.bin:$PATH"
 
           git config --global user.name nextstrain-bot
-          git config --global user.email 78992647+nextstrain-bot@users.noreply.github.com
+          git config --global user.email nextstrain-bot@users.noreply.github.com
 
           cd packages/web/
 
+          # Get the version from the previous step
           export AUSPICE_VERSION="${{ steps.check-auspice-version.outputs.AUSPICE_VERSION }}"
 
+          # Do the update
           yarn add auspice@${AUSPICE_VERSION}
 
+          # The PR will only have package.json and yarn.lock files
           git add package.json yarn.lock
           git commit -m "feat: update Auspice to version ${AUSPICE_VERSION}"
+
+          # Revert any other changes so that they don't get commited by the next step
+          git reset --hard
 
       - name: "Create Pull Request"
         id: create-pull-request
