### 2021-10-27 d289116
chore: cleanup unrelated files in GitHub Action, attempt 2
