+++ .github/workflows/Update-Auspice.yml
@@ -59,17 +59,19 @@ jobs:
           git config --global user.name nextstrain-bot
           git config --global user.email nextstrain-bot@users.noreply.github.com
 
-          cd packages/web/
+          pushd packages/web >/dev/null
 
-          # Get the version from the previous step
-          export AUSPICE_VERSION="${{ steps.check-auspice-version.outputs.AUSPICE_VERSION }}"
+            # Get the version from the previous step
+            export AUSPICE_VERSION="${{ steps.check-auspice-version.outputs.AUSPICE_VERSION }}"
 
-          # Do the update
-          yarn add auspice@${AUSPICE_VERSION}
+            # Do the update
+            yarn add auspice@${AUSPICE_VERSION}
 
-          # The PR will only have package.json and yarn.lock files
-          git add package.json yarn.lock
-          git commit -m "feat: update Auspice to version ${AUSPICE_VERSION}"
+            # The PR will only have package.json and yarn.lock files
+            git add package.json yarn.lock
+            git commit -m "feat: update Auspice to version ${AUSPICE_VERSION}"
+
+          popd >/dev/null
 
           # Revert any other changes so that they don't get commited by the next step
           git reset --hard
