### 2021-10-27 9aa6ed3
chore: cleanup unrelated files in GitHub Action, attempt 3
