### 2021-10-27 6f30a57
chore: fix action name in the gh command
