+++ .github/workflows/Update-Auspice.yml
@@ -99,7 +99,7 @@ jobs:
             This automated Pull Request is generated by GitHub Action [${{github.run_id}}](https://github.com/nextstrain/nextclade/actions/runs/${{github.run_id}}?check_suite_focus=true).
             Update can be re-triggered manually in Github Actions UI or with [GitHub CLI](https://github.com/cli/cli):
             ```
-            gh workflow run 'Update Auspice package'
+            gh workflow run 'Update Auspice'
             ```
 
           labels: |
