### 2021-10-27 5887b76
chore: add link to auspice changelog
