### 2021-10-27 e0e7470
chore: improve auspice bot pr name
