### 2021-10-27 0cc22e9
chore: remove version from auspice version bump bot
