### 2021-10-28 68a2902
fix: detect out-of-bounds access on array
