### 2021-10-28 c15ff3d
fix: only use sequence aa mutations for the current gene
