+++ packages/nextclade/src/tree/findPrivateNucMutations.cpp
@@ -1,3 +1,30 @@
+/**
+* This module implements search for *private* aminoacid mutations.
+*
+* Private mutations are the mutations in the query (user-provided) sequence relative to the parent node on the
+* reference tree.
+*
+* We have an array of sequence mutations relative to reference and a map of node mutations relative to reference.
+* We want to find private mutations. The following cases are possible:
+* |---|--------------------------------------------------------------|---------|------------------------------|
+* |   |                                                              |         |   Resulting private mutation |
+* |---|                         Case                                 | Private |------------------------------|
+* |   |                                                              |         |      From     |     To       |
+* |---|--------------------------------------------------------------|---------|---------------|--------------|
+* | 1 | mutation in sequence and in node, same query character       |   no    |      N/A      |     N/A      |
+* |---|--------------------------------------------------------------|---------|---------------|--------------|
+* | 2 | mutation in sequence and in node, but not the same character |   yes   |    node.qry   |   seq.qry    |
+* |---|--------------------------------------------------------------|---------|---------------|--------------|
+* | 3 | mutation in sequence but not in node                         |   yes   |    seq.ref    |   seq.qry    |
+* |---|--------------------------------------------------------------|---------|---------------|--------------|
+* | 4 | mutation in node, but not in sequence                        |   yes   |    node.qry   |   node.ref   |
+* |   | (mutation in sequence that reverts the character to ref seq) |         |               |              |
+* |---|--------------------------------------------------------------|---------|---------------|--------------|
+*
+* At this point sequence have not yet become a new node on the tree, but is described by the results of the previous
+* analysis steps.
+*
+*/
 #include "findPrivateNucMutations.h"
 
 #include <fmt/format.h>
@@ -14,209 +41,216 @@
 #include "../utils/filter.h"
 #include "../utils/mapFind.h"
 
+
 namespace Nextclade {
-  template<typename Letter>
-  bool isSequencedGeneric(int pos, const AnalysisResult& seq, LetterTag<Letter>);
+  namespace {
 
-  template<>
-  bool isSequencedGeneric(int pos, const AnalysisResult& seq, LetterTag<Nucleotide>) {
-    return isSequenced(pos, seq);
-  }
+    /** Generic version of isSequenced() function */
+    template<typename Letter>
+    bool isSequencedGeneric(int pos, const AnalysisResult& seq, LetterTag<Letter>);
 
-  template<>
-  bool isSequencedGeneric(int /*pos*/, const AnalysisResult& /*seq*/, LetterTag<Aminoacid>) {
-    // For aminoacid sequences there is no concept of being not sequences currently. They are always sequenced.
-    return true;
-  }
-
-  template<typename Letter>
-  void processSeqSubstitutions(                                               //
-    /* in */ const std::map<int, Letter>& nodeMutMap,                         //
-    /* in */ const std::vector<Substitution<Letter>>& substitutions,          //
-    /* inout */ std::vector<SubstitutionSimple<Letter>>& privateSubstitutions,//
-    /* inout */ std::set<int>& seqPositionsCovered                            //
-  ) {
-    for (const auto& seqMut : substitutions) {
-      const auto& pos = seqMut.pos;
-      const auto& nodeQueryChar = mapFind(nodeMutMap, pos);
-      seqPositionsCovered.insert(pos);
-
-      if (!nodeQueryChar) {
-        // Case 3: Mutation in sequence but not in node, i.e. a newly occurred mutation.
-        // Action: Add the sequence mutation itself.
[…]
