### 2021-10-28 79d0e4a
chore: restore commented script blocks
