### 2021-10-28 e3be0b9
fix: avoid addition of an empty array if there are no private mutations
