### 2021-10-28 98f2471
docs: add citation suggestion to docs/index.rst
