### 2021-10-28 afe767c
chore(docs): add line block to author list
