### 2021-10-29 27f2528
fix: ensure private nuc mutations are on the tree again
The mutations object was incorrectly overwritten entirely with and object containing amino acids only. So let's move the nuc and aa code closer together and construct one object instead of two. No overwrites this way.
