+++ packages/nextclade/src/io/formatMutation.cpp
@@ -36,6 +36,12 @@ namespace Nextclade {
     return fmt::format("{}{}{}", nucToString(mut.ref), positionOneBased, nucToString(mut.qry));
   }
 
+  std::string formatDeletionSimple(const NucleotideDeletionSimple& del) {
+    // NOTE: by convention, in bioinformatics, nucleotides are numbered starting from 1, however our arrays are 0-based
+    const auto positionOneBased = del.pos + 1;
+    return fmt::format("{}{}{}", nucToString(del.ref), positionOneBased, "-");
+  }
+
   std::string formatAminoacidMutationSimpleWithoutGene(const AminoacidSubstitutionSimple& mut) {
     // NOTE: by convention, in bioinformatics, nucleotides are numbered starting from 1, however our arrays are 0-based
     const auto positionOneBased = mut.pos + 1;
