+++ packages/nextclade/src/tree/TreeNode.cpp
@@ -353,19 +353,27 @@ namespace Nextclade {
       }
     }
 
-    void setBranchAttrAaMutations(const std::map<std::string, PrivateAminoacidMutations>& aaMutations) {
+    void setBranchAttrAaMutations(                                       //
+      const PrivateNucleotideMutations& nucMutations,                    //
+      const std::map<std::string, PrivateAminoacidMutations>& aaMutations//
+    ) {
       auto mutObj = json::object();
 
-      if (aaMutations.empty()) {
-        return;
+
+      for (const auto& mut : nucMutations.privateSubstitutions) {
+        mutObj["nuc"].push_back(formatMutationSimple(mut));
+      }
+
+      for (const auto& mut : nucMutations.privateDeletions) {
+        mutObj["nuc"].push_back(formatDeletionSimple(mut));
       }
 
       for (const auto& [geneName, aaMutationsForGene] : aaMutations) {
-        const auto& privateSubstitutions = aaMutationsForGene.privateSubstitutions;
-        const auto& privateDeletions = aaMutationsForGene.privateDeletions;
-        const int totalMutations = safe_cast<int>(privateSubstitutions.size() + privateDeletions.size());
+        const auto& privateAaSubstitutions = aaMutationsForGene.privateSubstitutions;
+        const auto& privateAaDeletions = aaMutationsForGene.privateDeletions;
+        const int totalPrivateAaMutations = safe_cast<int>(privateAaSubstitutions.size() + privateAaDeletions.size());
 
-        if (totalMutations == 0) {
+        if (totalPrivateAaMutations == 0) {
           continue;
         }
 
@@ -373,11 +381,11 @@ namespace Nextclade {
           mutObj[geneName] = json::array();
         }
 
-        for (const auto& mut : privateSubstitutions) {
+        for (const auto& mut : privateAaSubstitutions) {
           mutObj[geneName].push_back(formatAminoacidMutationSimpleWithoutGene(mut));
         }
 
-        for (const auto& mut : privateDeletions) {
+        for (const auto& mut : privateAaDeletions) {
           mutObj[geneName].push_back(formatAminoacidDeletionSimpleWithoutGene(mut));
         }
       }
@@ -547,12 +555,9 @@ namespace Nextclade {
     pimpl->setNucleotideMutationsEmpty();
   }
 
-  void TreeNode::setBranchAttrNucMutations(const std::vector<NucleotideSubstitutionSimple>& mutations) {
-    pimpl->setBranchAttrNucMutations(mutations);
-  }
-
-  void TreeNode::setBranchAttrAaMutations(const std::map<std::string, PrivateAminoacidMutations>& aaMutations) {
-    pimpl->setBranchAttrAaMutations(aaMutations);
+  void TreeNode::setBranchAttrMutations(const PrivateNucleotideMutations& nucMutations,
+    const std::map<std::string, PrivateAminoacidMutations>& aaMutations) {
+    pimpl->setBranchAttrAaMutations(nucMutations, aaMutations);
   }
 
   std::optional<double> TreeNode::divergence() const {
