### 2021-10-29 b1bb20c
refactor: remove unused functions
