+++ packages/nextclade/src/io/serializeResults.h
@@ -8,6 +8,15 @@ namespace Nextclade {
 
   using json = nlohmann::ordered_json;
 
+  template<typename Container>
+  inline json serializeArray(const Container& container) {
+    auto j = json::array();
+    for (const auto& elem : container) {
+      j.template emplace_back(elem);
+    }
+    return j;
+  }
+
   template<typename Container, typename Serializer>
   inline json serializeArray(const Container& container, Serializer serializer) {
     auto j = json::array();
