### 2021-10-29 3b0a905
fix: find private mutations in all genes, not just the ones on the tree
