### 2021-10-29 4dba002
docs(citation): prettify citation on main page of docs
