### 2021-10-29 63deb50
docs(citation): serve bibtex file
