+++ packages/web/public/citation.bib
@@ -1,0 +1,14 @@
+@misc{ivan_aksamentov_2021_5607694,
+  author       = {Ivan Aksamentov and
+                  Cornelius Roemer and
+                  Emma B Hodcroft and
+                  Richard A Neher},
+  title        = {{Nextclade: clade assignment, mutation calling and
+                   quality control for viral genomes}},
+  month        = oct,
+  year         = 2021,
+  note         = {available at https://clades.nextstrain.org},
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.5607694},
+  url          = {https://doi.org/10.5281/zenodo.5607694}
+}
