### 2021-10-29 3e56e04
docs(citation): add bibtex links
