+++ docs/index.rst
@@ -26,6 +26,8 @@ Nextclade is a part of `Nextstrain <https://nextstrain.org>`_, an open-source pr
 
       Ivan Aksamentov, Cornelius Roemer, Emma B Hodcroft, & Richard A Neher. (2021). Nextclade: clade assignment, mutation calling and quality control for viral genomes. https://doi.org/10.5281/zenodo.5607694
 
+    `bibtex <https://zenodo.org/record/5607694/export/hx>`_
+
   - where possible, provide a link to Nextclade Web:
 
       https://clades.nextstrain.org
