+++ README.md
@@ -205,6 +205,7 @@ If you use results obtained with Nextclade in a publication, please
 
     > Ivan Aksamentov, Cornelius Roemer, Emma B Hodcroft, & Richard A Neher. (2021). Nextclade: clade assignment, mutation calling and quality control for viral genomes. https://doi.org/10.5281/zenodo.5607694
 
+   ([bibtex](https://zenodo.org/record/5607694/export/hx))
 
  - where possible, provide a link to Nextclade Web:
 
