### 2021-10-29 22aadef
feat: filter out private substitutions to nucleotide N and aminoacid X
