### 2021-10-29 e413193
refactor: cleanup old debugging code
