### 2021-11-01 7f3213e
chore(ci): add staging builds for cli on ci server
