### 2021-11-01 c4be26a
docs: extend deprecation notice, mention files that remain
