### 2021-11-01 #3773 @ivan-aksamentov
@kevinlibuit
> Also noticed that the Nextalign quick example suggests grabbing data files from the defunct data/sars-cov-2 directory.

Good point! I thought we might update the message to make it clearer that we keep the Nextalign files in place and only Nextclade files will be removed.
https://github.com/nextstrain/nextclade/blob/b953cfb5fb4002c2d55d07fc16762be949bcbdfc/data/README.md

This change is still in progress and will be finalized on November 10, 2021 with the PR https://github.com/nextstrain/nextclade/pull/597
