+++ packages/nextalign/src/translate/translateGenes.h
@@ -6,11 +6,11 @@
 #include <map>
 #include <vector>
 
-PeptidesInternal translateGenes(                       //
-  const NucleotideSequence& query,                     //
-  const NucleotideSequence& ref,                       //
+PeptidesInternal translateGenes(                               //
+  const NucleotideSequence& query,                             //
+  const NucleotideSequence& ref,                               //
   const std::map<std::string, RefPeptideInternal>& refPeptides,//
-  const GeneMap& geneMap,                              //
-  const std::vector<int>& gapOpenCloseAA,              //
-  const NextalignOptions& options                      //
+  const GeneMap& geneMap,                                      //
+  const std::vector<int>& gapOpenCloseAA,                      //
+  const NextalignOptions& options                              //
 );
