+++ packages/nextclade/src/__tests__/getPcrPrimerChanges.test.cpp
@@ -1,3 +1,5 @@
+#include "../analyze/getPcrPrimerChanges.h"
+
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
@@ -5,15 +7,14 @@
 
 #include "../../include/nextclade/nextclade.h"
 #include "../../include/nextclade/private/nextclade_private.h"
-#include "../analyze/getPcrPrimerChanges.h"
 
 #define EXPECT_ARR_EQ(expected, actual) ASSERT_THAT(actual, ::testing::ElementsAreArray(expected))
 
 namespace {
+  using ::Range;
   using ::Nextclade::NucleotideLocation;
   using ::Nextclade::NucleotideSubstitution;
   using ::Nextclade::PcrPrimer;
-  using ::Range;
   using ::Nextclade::shouldReportPrimerMutation;
 
   NucleotideSubstitution makeMutation(int pos, char query) {
