### 2021-11-01 aa00607
chore: bring back csv outputs validation test
