+++ scripts/build_locally.sh
@@ -845,9 +845,9 @@ pushd "${PROJECT_ROOT_DIR}" > /dev/null
      eval "${GDB}" ${NEXTCLADE_CLI} ${DEV_NEXTCLADE_CLI_OPTIONS} || cd .
     fi
 
-#      print 25 "Validate CSV/TSV outputs";
-#      ./scripts/csv-validator.sh
-#      ./scripts/csvlint.sh
+    print 25 "Validate CSV/TSV outputs";
+    ./scripts/csv-validator.sh
+    ./scripts/csvlint.sh
     fi
 
   print 22 "Done";
