### 2021-11-04 e7e0b4c
fix: count private del ranges instead of individual dels in qc
