### 2021-11-04 6f3847c
fix: terminate the last deletion range
