### 2021-11-08 #3773 @kevinlibuit
Thanks for the info @rneher! Really helpful to see those metrics outright. Is there an obvious rationale behind those discrepancies in speed, e.g. is it the underlying approach or something done on the compute/programming side of things that contribute most to the efficiency.

^^ Asking more so out of curiosity at this point. Everything provided meets my understanding of the acceptance criteria for JOSS and I give my full recommendation for acceptance of this manuscript.
