### 2021-11-10 fe35b10
fix: avoid race between data fetch and read when url params are used
When datasets feature was added, a data race was introduced due to lack of synchronization between asynchronous data fetching from locations provided by URL params and reading this data before launching the algorithm.

The order of operation was as follows:
1. Retrieve URLs from URL params
2. Start fetching from these urls
3. Save fetched data in redux store upon arrival
4. Read data from redux store. If it's empty, load default data from a dataset.
5. Do some other operations.
6. Launch the algorithm.

However, step 4 was no
[…]
