### 2021-11-16 297f44f
chore: download dataset as a part of build script
Now that the data is no longer in the repo, we need to download a dataset in order to be able to run Nextclade during routine development. This adds `nextclade dataset get` command to the build script.
