### 2021-11-16 a979442
feat: add example alignment with bio-rust
