+++ packages/nextalign_rs/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "nextalign"
+description = "Nucleotide sequence alignment and translation"
+version = "2.0.0"
+edition = "2021"
+license = "MIT"
+publish = false
+
+[dependencies]
+bio = "0.39.0"
