### 2021-11-17 6f307df
feat: read sequences from files
