### 2021-11-17 8a80676
perf: add custom allocator and perf tweaks
