### 2021-11-18 479906c
fix: ensure frame shifts in codon coordinates have size of at least 1
