+++ packages/nextalign/src/translate/detectFrameShifts.cpp
@@ -199,13 +199,15 @@ std::vector<InternalFrameShiftResultWithMask> translateFrameShifts(//
     Range nucAbsAln = nucRelAln + geneStartAln;
     Range nucAbsRef = coordMap.alnToRef(nucAbsAln);
     Range nucRelRef = nucAbsRef - geneStartRef;
-    Range codon = nucRelRef / 3;
+    Range codon = (nucRelRef / 3);
+    codon.ensureSizeOfAtLeastOne();
 
     Range maskNucRelAln = findMask(query, nucRelAln);
     Range maskNucAbsAln = maskNucRelAln + geneStartAln;
     Range maskNucAbsRef = coordMap.alnToRef(maskNucAbsAln);
     Range maskNucRelRef = maskNucAbsRef - geneStartRef;
     Range maskCodon = maskNucRelRef / 3;
+    maskCodon.ensureSizeOfAtLeastOne();
 
     FrameShiftContext gapsLeading{
       .codon{
