### 2021-11-18 bd0c384
feat: add invariants around ranges in frame shift detector code
This is helpful for debugging
