### 2021-11-18 9874559
fix: ensure that the frame shift codon range calculation is greedy
We want to count partially shifted codons and as a particular case, small frame shifts that cover only a portion of a codon. 

So for example, now if 4 nucs are covered, it will be 2 codons, not 1. Similarly, if 2 nucs are covered, this will be 1 codon, up from 0, which was the initial big to fix.

In this commit the test expectations are also adjusted to match.
