### 2021-11-19 4996e54
feat: add debug trace statements in the alignment code
