### 2021-11-19 d37b0f8
feat: skip seed match for aa and infer params from aligned full seqs
