+++ packages/nextalign/src/nextalign.cpp
@@ -1,6 +1,8 @@
 #include <nextalign/nextalign.h>
 #include <utils/concat_move.h>
 
+#include <algorithm>
+
 #include "align/alignPairwise.h"
 #include "align/getGapOpenCloseScores.h"
 #include "strip/stripInsertions.h"
@@ -47,18 +49,26 @@ NextalignResultInternal nextalignInternal(const NucleotideSequence& query, const
     throw ErrorNonFatal(*alignmentStatus.error);
   }
 
+  const auto stripped = stripInsertions(alignmentStatus.result->ref, alignmentStatus.result->query);
+  const auto refStripped = removeGaps(ref);
+
   std::vector<PeptideInternal> queryPeptides;
   Warnings warnings;
   if (!geneMap.empty()) {
+    const auto querySize = safe_cast<int>(alignmentStatus.result->query.size());
+    const auto queryStrippedSize = safe_cast<int>(stripped.queryStripped.size());
+    const auto refSize = safe_cast<int>(alignmentStatus.result->ref.size());
+    const auto refStrippedSize = safe_cast<int>(refStripped.size());
+
+    const int bandWidth = safe_cast<int>(std::max(querySize - queryStrippedSize, refSize - refStrippedSize) / 3 + 3);
+    const int shift = bandWidth / 2;
+
     auto peptidesInternal = translateGenes(alignmentStatus.result->query, alignmentStatus.result->ref, refPeptides,
-      geneMap, gapOpenCloseAA, options);
+      geneMap, gapOpenCloseAA, options, bandWidth, shift);
     concat_move(peptidesInternal.queryPeptides, queryPeptides);
     concat_move(peptidesInternal.warnings.global, warnings.global);
     concat_move(peptidesInternal.warnings.inGenes, warnings.inGenes);
   }
-
-  const auto stripped = stripInsertions(alignmentStatus.result->ref, alignmentStatus.result->query);
-  const auto refStripped = removeGaps(ref);
 
   return NextalignResultInternal{
     .query = stripped.queryStripped,
