### 2021-11-19 be0af1b
feat: deduce peptide band width per gene
