### 2021-11-19 68e6d02
chore: reenable debug trace
