### 2021-11-19 4f2e2a2
refactor: default-initialize missing fields
