+++ packages/nextclade/src/analyze/findNucChanges.cpp
@@ -45,6 +45,7 @@ namespace Nextclade {
           .qry = d,
           .pcrPrimersChanged = {},
           .aaSubstitutions = {},
+          .aaDeletions = {},
         });
       } else if (isGap(d) && !beforeAlignment) {
         if (!nDel) {
