### 2021-11-22 9a53c70
refactor: move alignment param caculation into a separate function
To make it testable
