### 2021-11-22 e444443
refactor: move alignment param calculation into a separate file
To make it testable
