### 2021-11-22 2035136
fix: missing division in aa alignment shift calculation
