### 2021-11-22 a57cc6e
fix: limit search of trailing gaps up to the first leading gap
To avoid double-counting
