### 2021-11-22 9c7dc8c
fix: off-by-1 issue in trailing gap number calculation
