### 2021-11-23 f6e0405
feat: enable or disable debug tracing using build-time flag
