+++ packages/nextalign/src/utils/debug_trace.h
@@ -3,8 +3,22 @@
 #include <array>
 #include <vector>
 
+#define NOOP ({ (void) 0; })
 
-#define debug_trace(format_string, ...) fmt::print(format_string, __VA_ARGS__)
+// If `ENABLE_DEBUG_TRACE` is set to `1`, the `debug_trace()` statements throughout the codebase
+// will be printing messages to the console. Useful for algorithm development.
+// If disabled, `debug_trace()` evaluates to a no-op, a dummy statement which should be
+// optimized-away by the compiler.
+// Enabling `debug_trace()` incurs significant performance penalty. Do not enable in production!
+#ifndef ENABLE_DEBUG_TRACE
+#define ENABLE_DEBUG_TRACE 0
+#endif
+
+#if ENABLE_DEBUG_TRACE == 1
+#define debug_trace(format_string, ...) ({ fmt::print(format_string, __VA_ARGS__); })
+#else
+#define debug_trace(format_string, ...) NOOP
+#endif
 
 template<typename T>
 struct fmt::formatter<std::vector<T>> {
