### 2021-11-23 e66d11d
docs: clarify comments about debug tracing
