### 2021-11-23 9a2d4fe
feat: copy node attrs from ref tree into new nodes and into results json
