### 2021-11-23 #3773 @kellyrowland
Thanks for the ping on this, @rneher . Since we have two reviewers who have recommended the submission for acceptance, I'll proceed with the checkout process.

(That is, I will check off the review boxes for DavidNickle and remove EduanWilkinson as a reviewer.)

Could you please create a new tagged software release (if the software has changed in this review process) and archive the release (on Zenodo, figshare, etc.) and then post the version number and archive DOI here in this review issue?
