### 2021-11-23 #3773 @kellyrowland
I have updated the top post in this issue as per my above note.

@EduanWilkinson thank you for your interest here - we're going to proceed with the paper acceptance since we have positive recommendations from two reviewers at this point. I do hope that you'll consider reviewing for JOSS in the future.
