### 2021-11-23 e404f58
feat: write dynamic tree node attributes into csv and tsv otuputs
