+++ packages/nextclade/src/io/CsvWriter.cpp
@@ -11,8 +11,10 @@
 #include <vector>
 
 #include "../utils/at.h"
+#include "../utils/concat.h"
 #include "../utils/contains.h"
 #include "../utils/contract.h"
+#include "../utils/eraseDuplicates.h"
 #include "../utils/safe_cast.h"
 #include "formatMutation.h"
 #include "formatQcStatus.h"
@@ -20,87 +22,92 @@
 namespace Nextclade {
 
   namespace {
-    constexpr auto COLUMN_NAMES = frozen::make_map<frozen::string, int>({
-      {"seqName", 0},
-      {"clade", 1},
 
-      {"qc.overallScore", 2},
-      {"qc.overallStatus", 3},
+#pragma clang diagnostic push
+#pragma ide diagnostic ignored "cppcoreguidelines-avoid-magic-numbers"
+    inline std::vector<std::string> getDefaultColumnNames() {
+      return std::vector<std::string>{
+        "seqName",
+        "clade",
 
-      {"totalSubstitutions", 4},
-      {"totalDeletions", 5},
-      {"totalInsertions", 6},
-      {"totalFrameShifts", 7},
-      {"totalAminoacidSubstitutions", 8},
-      {"totalAminoacidDeletions", 9},
-      {"totalMissing", 10},
-      {"totalNonACGTNs", 11},
-      {"totalPcrPrimerChanges", 12},
+        "qc.overallScore",
+        "qc.overallStatus",
 
-      {"substitutions", 13},
-      {"deletions", 14},
-      {"insertions", 15},
-      {"frameShifts", 16},
-      {"aaSubstitutions", 17},
-      {"aaDeletions", 18},
-      {"missing", 19},
-      {"nonACGTNs", 20},
-      {"pcrPrimerChanges", 21},
+        "totalSubstitutions",
+        "totalDeletions",
+        "totalInsertions",
+        "totalFrameShifts",
+        "totalAminoacidSubstitutions",
+        "totalAminoacidDeletions",
+        "totalMissing",
+        "totalNonACGTNs",
+        "totalPcrPrimerChanges",
 
-      {"alignmentScore", 22},
-      {"alignmentStart", 23},
-      {"alignmentEnd", 24},
+        "substitutions",
+        "deletions",
+        "insertions",
+        "frameShifts",
+        "aaSubstitutions",
+        "aaDeletions",
+        "missing",
+        "nonACGTNs",
+        "pcrPrimerChanges",
 
-      {"qc.missingData.missingDataThreshold", 25},
-      {"qc.missingData.score", 26},
-      {"qc.missingData.status", 27},
-      {"qc.missingData.totalMissing", 28},
+        "alignmentScore",
+        "alignmentStart",
+        "alignmentEnd",
 
-      {"qc.mixedSites.mixedSitesThreshold", 29},
-      {"qc.mixedSites.score", 30},
-      {"qc.mixedSites.status", 31},
-      {"qc.mixedSites.totalMixedSites", 32},
+        "qc.missingData.missingDataThreshold",
+        "qc.missingData.score",
+        "qc.missingData.status",
+        "qc.missingData.totalMissing",
 
-      {"qc.privateMutations.cutoff", 33},
-      {"qc.privateMutations.excess", 34},
-      {"qc.privateMutations.score", 35},
-      {"qc.privateMutations.status", 36},
-      {"qc.privateMutations.total", 37},
+        "qc.mixedSites.mixedSitesThreshold",
+        "qc.mixedSites.score",
+        "qc.mixedSites.status",
+        "qc.mixedSites.totalMixedSites",
 
-      {"qc.snpClusters.clusteredSNPs", 38},
-      {"qc.snpClusters.score", 39},
-      {"qc.snpClusters.status", 40},
-      {"qc.snpClusters.totalSNPs", 41},
[…]
