+++ packages/nextclade_wasm/src/main.cpp
@@ -13,6 +13,18 @@
 #include "../../packages/nextclade/src/tree/treePostprocess.h"
 #include "../../packages/nextclade/src/tree/treePreprocess.h"
 
+// FIXME: node attributes should be dynamic. Should they come from a dataset? From the tree `.meta` perhaps?
+std::vector<std::string> getCustomNodeAttrKeys() {
+  return {
+    "GISAID_clade",
+    "Nextstrain_clade",
+    "WHO_name",
+    "clade",
+    "clade_shortname",
+    "pango_lineage",
+  };
+}
+
 std::string getExceptionMessage(std::intptr_t exceptionPtr) {// NOLINT(misc-unused-parameters)
   // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast,cppcoreguidelines-init-variables,performance-no-int-to-ptr)
   const std::exception* e = reinterpret_cast<std::runtime_error*>(exceptionPtr);
@@ -155,7 +167,8 @@ public:
         state.pcrPrimers,                    //
         state.qcRulesConfig,                 //
         state.tree,                          //
-        state.nextalignOptions               //
+        state.nextalignOptions,              //
+        getCustomNodeAttrKeys()              //
       );
 
       warnings.global = merge(warnings.global, result.warnings.global);
@@ -244,7 +257,8 @@ std::string treeFinalize(const std::string& treeStr, const std::string& refStr, 
 std::string serializeToCsv(const std::string& analysisResultsStr, const std::string& delimiter) {
   const auto analysisResults = wrappedParseAnalysisResults(analysisResultsStr, "'serializeToCsv'");
   std::stringstream outputCsvStream;
-  auto csv = Nextclade::createCsvWriter(Nextclade::CsvWriterOptions{.delimiter = delimiter[0]});
+  auto csv =
+    Nextclade::createCsvWriter(Nextclade::CsvWriterOptions{.delimiter = delimiter[0]}, getCustomNodeAttrKeys());
   for (const auto& result : analysisResults.results) {
     csv->addRow(result);
   }

[…]
