### 2021-11-23 nextstrain/nextclade#584 @ivan-aksamentov
We discussed internally and decided that we keep the files in `data/` for Nextalign, so this remains the recommended way to get data for Nextalign. For Nextclade we recommend using datasets. I will close this for now.

In the future we may (or may not) unify this. Adding networking makes binaries bigger and slower and we want Nextalign to be as fast as possible. Plus, in `data/` we have more pathogens than in the Nextclade datasets. And we currently have a notion of Nextalign dataset. This is something to think about, but does not seem to be a priority currently.
