### 2021-11-24 152f8be
fix: tentative fix for a SIGSEGV crash in coordinate mapper
