### 2021-11-24 274cb31
fix: tentative fix of a sigsegv crash in coordinate mapper
This uses the new implementation of funcions that create the maps. These functions don't have a faulty `revCoordMap.back()` call.
