### 2021-11-25 42e84e5
feat: use dynamic list of clade-like node attrs from the tree json
An example extension of tree JSON is below. Paste it as an additional property into the `meta` object.

```json
    "extensions": {
      "nextclade": {
        "clade_node_attrs_keys": [
          "GISAID_clade",
          "Nextstrain_clade",
          "WHO_name",
          "clade",
          "clade_shortname",
          "pango_lineage"
        ]
      }
    }
```
