+++ packages/nextclade/include/nextclade/nextclade.h
@@ -159,7 +159,6 @@ namespace Nextclade {
     GeneMap geneMap;
     QcConfig qcRulesConfig;
     NextalignOptions nextalignOptions;
-    std::vector<std::string> customNodeAttrKeys;
   };
 
   template<typename Letter>
@@ -442,6 +441,7 @@ namespace Nextclade {
     std::string schemaVersion;
     std::string nextcladeVersion;
     std::uint64_t timestamp;
+    std::vector<std::string> cladeNodeAttrKeys;
     std::vector<Nextclade::AnalysisResult> results;
   };
 
@@ -470,6 +470,8 @@ namespace Nextclade {
 
   public:
     explicit NextcladeAlgorithm(const NextcladeOptions& options);
+
+    std::vector<std::string> getCladeNodeAttrKeys() const;
 
     NextcladeResult run(const std::string& seqName, const NucleotideSequence& seq);
 
@@ -550,6 +552,8 @@ namespace Nextclade {
 
     TreeNode root() const;
 
+    std::vector<std::string> getCladeNodeAttrKeys() const;
+
     void addMetadata();
 
     double tmpMaxDivergence() const;
@@ -592,7 +596,7 @@ namespace Nextclade {
 
   std::string serializeResultToString(const AnalysisResult& result);
 
-  std::string serializeResults(const std::vector<AnalysisResult>& results);
+  std::string serializeResults(const AnalysisResults& results);
 
   std::string formatRange(const Range& range);
 
