+++ packages/nextclade/src/nextclade.cpp
@@ -177,24 +177,28 @@ namespace Nextclade {
       treePreprocess(tree, opt.ref, refPeptides);
     }
 
+    std::vector<std::string> getCladeNodeAttrKeys() const {
+      return tree.getCladeNodeAttrKeys();
+    }
+
     NextcladeResult run(const std::string& seqName, const NucleotideSequence& query) {
       const auto& ref = options.ref;
       const auto& pcrPrimers = options.pcrPrimers;
       const auto& geneMap = options.geneMap;
       const auto& qcRulesConfig = options.qcRulesConfig;
 
-      return analyzeOneSequence(  //
-        seqName,                  //
-        ref,                      //
-        query,                    //
-        refPeptides,              //
-        refPeptidesArr,           //
-        geneMap,                  //
-        pcrPrimers,               //
-        qcRulesConfig,            //
-        tree,                     //
-        options.nextalignOptions, //
-        options.customNodeAttrKeys//
+      return analyzeOneSequence(   //
+        seqName,                   //
+        ref,                       //
+        query,                     //
+        refPeptides,               //
+        refPeptidesArr,            //
+        geneMap,                   //
+        pcrPrimers,                //
+        qcRulesConfig,             //
+        tree,                      //
+        options.nextalignOptions,  //
+        tree.getCladeNodeAttrKeys()//
       );
     }
 
@@ -213,6 +217,10 @@ namespace Nextclade {
       : pimpl(std::make_unique<NextcladeAlgorithmImpl>(options)) {}
 
   NextcladeAlgorithm::~NextcladeAlgorithm() {}// NOLINT(modernize-use-equals-default)
+
+  std::vector<std::string> NextcladeAlgorithm::getCladeNodeAttrKeys() const {
+    return pimpl->getCladeNodeAttrKeys();
+  }
 
   NextcladeResult NextcladeAlgorithm::run(const std::string& seqName, const NucleotideSequence& seq) {
     return pimpl->run(seqName, seq);
