+++ packages/nextclade_cli/src/algorithm/runNextclade.cpp
@@ -14,6 +14,8 @@
 
 #include "../io/Logger.h"
 #include "../io/parseRefFastaFile.h"
+#include "../utils/datetime.h"
+#include "../utils/safe_cast.h"
 
 
 namespace Nextclade {
@@ -54,26 +56,6 @@ namespace Nextclade {
     const auto &ref = refData.seq;
     const auto &refName = refData.name;
 
-    // FIXME: node attributes should be dynamic. Should they come from a dataset? From the tree `.meta` perhaps?
-    const std::vector<std::string> customNodeAttrKeys = {
-      "GISAID_clade",
-      "Nextstrain_clade",
-      "WHO_name",
-      "clade",
-      "clade_shortname",
-      "pango_lineage",
-    };
-
-    std::unique_ptr<Nextclade::CsvWriterAbstract> csv;
-    if (outputCsvStream) {
-      csv = createCsvWriter(CsvWriterOptions{.delimiter = ';'}, customNodeAttrKeys);
-    }
-
-    std::unique_ptr<Nextclade::CsvWriterAbstract> tsv;
-    if (outputTsvStream) {
-      tsv = createCsvWriter(CsvWriterOptions{.delimiter = '\t'}, customNodeAttrKeys);
-    }
-
     const NextcladeOptions options = {
       .ref = ref,
       .treeString = treeString,
@@ -81,10 +63,20 @@ namespace Nextclade {
       .geneMap = geneMap,
       .qcRulesConfig = qcRulesConfig,
       .nextalignOptions = nextalignOptions,
-      .customNodeAttrKeys = customNodeAttrKeys,
     };
 
     NextcladeAlgorithm nextclade{options};
+    const auto cladeNodeAttrKeys = nextclade.getCladeNodeAttrKeys();
+
+    std::unique_ptr<Nextclade::CsvWriterAbstract> csv;
+    if (outputCsvStream) {
+      csv = createCsvWriter(CsvWriterOptions{.delimiter = ';'}, cladeNodeAttrKeys);
+    }
+
+    std::unique_ptr<Nextclade::CsvWriterAbstract> tsv;
+    if (outputTsvStream) {
+      tsv = createCsvWriter(CsvWriterOptions{.delimiter = '\t'}, cladeNodeAttrKeys);
+    }
 
     // TODO(perf): consider using a thread-safe queue instead of a vector,
     //  or restructuring code to avoid concurrent access entirely
@@ -237,7 +229,13 @@ namespace Nextclade {
       std::vector<AnalysisResult> results{resultsConcurrent.cbegin(), resultsConcurrent.cend()};
 
       if (outputJsonStream) {
-        *outputJsonStream << serializeResults(results);
+        *outputJsonStream << serializeResults(AnalysisResults{
+          .schemaVersion = Nextclade::getAnalysisResultsJsonSchemaVersion(),
+          .nextcladeVersion = Nextclade::getVersion(),
+          .timestamp = safe_cast<uint64_t>(getTimestampNow()),
+          .cladeNodeAttrKeys = cladeNodeAttrKeys,
+          .results = results,
+        });
       }
 
       if (outputTreeStream) {
