+++ packages/nextclade_cli/src/utils/datetime.h
@@ -1,0 +1,6 @@
+#include <chrono>
+
+auto getTimestampNow() {
+  const auto p1 = std::chrono::system_clock::now();
+  return std::chrono::duration_cast<std::chrono::seconds>(p1.time_since_epoch()).count();
+}
