+++ packages/nextclade/src/io/parseAnalysisResults.cpp
@@ -392,6 +392,7 @@ namespace Nextclade {
         .schemaVersion = at(j, "schemaVersion"),
         .nextcladeVersion = at(j, "nextcladeVersion"),
         .timestamp = at(j, "timestamp"),
+        .cladeNodeAttrKeys = at(j, "cladeNodeAttrKeys"),
         .results = parseArray<AnalysisResult>(j, "results", parseAnalysisResult),
       };
     } catch (const std::exception& e) {

[…]
