### 2021-11-25 #3773 @rneher
Hi @kellyrowland
I made a DOI for version 1.4.5 at Zenodo:

Version 1.4.5
10.5281/zenodo.5726681

https://zenodo.org/record/5726681

We'll have automated registration with zenodo for future releases, but weren't quite ready to make a new release just now.
