### 2021-11-25 dd73c0d
feat: show the logo conditionally depending on data provenance
