### 2021-11-25 02bb67d
fix: adjust for the fact that data provenance is an array
