### 2021-11-25 2ec4339
fix: adjust for meta properties name case changes in auspice
