### 2021-11-27 7707be7
chore: release cli 1.5.0 and web 1.8.0
