### 2021-11-27 b4e0275
fix: typo in Changelog
