+++ CHANGELOG.md
@@ -8,7 +8,7 @@ We improved the algorithm for peptide alignment. Instead of performing seed matc
 
 Nextclade previously did not account for the last codon in a frame shift if that codon was covered by the shift only partially. In this version we count the partial codons. This solves an issue with empty frame shift codon ranges being reported in rare cases. This change may result in some of the frame shifts to be longer by 1 codon in the new version compared to previous versions of Nextclade. The nucleotide length of frame shifts stays the same.
 
-None of the ignored frame shift ranges in the QC configurations of the exising dataset are affected byt this change. But if you use a custom QC configuration, some of the frame shifts in the list of ignored frame shifts might need to be be adjusted.
+None of the ignored frame shift ranges in the QC configurations of the exising dataset are affected by this change. But if you use a custom QC configuration, some of the frame shifts in the list of ignored frame shifts might need to be be adjusted.
 
 ### [Fix] Fix crashes with Nextclade CLI on macOS
 
