### 2021-11-28 fa49628
fix: avoid potential dereference of nullopt
If there is no open range by the end of the loop, then `begin` might be `nullopt`. Then it's illegal to dereference it in the body of this conditional. I add an additional check to see if `begin` contains a value.
