### 2021-11-28 7648924
fix: avoid adding an empty range
And while we are here, why don't we also check if the range is not empty before pushing it
