### 2021-11-28 7067acf
refactor: use min instead of a complex ternary expression
