### 2021-11-28 e653ad1
refactor: use existing signed integer variable instead of .size()
