+++ packages/nextalign/src/align/alignPairwise.cpp
@@ -375,12 +375,6 @@ AlignmentStatus<Letter> backTrace(const Sequence<Letter>& query, const Sequence<
     lastIndexByShift[i] = std::min(rowLength - 1, querySize + is);
 
     if (lastIndexByShift[i] >= 0 && lastIndexByShift[i] < scoresSize) {
-      // invariant_greater(lastIndexByShift[i], 0);
-      // invariant_less(lastIndexByShift[i], scores.num_cols());
-      // debug_trace(
-      //   "backtrace: i={:}, is={:}, lastIndexByShift={:}, bestScore={:}, score={:}\n",
-      //   i, is, lastIndexByShift[i], bestScore, scores(i, lastIndexByShift[i]));
-
       lastScoreByShift[i] = scores(i, lastIndexByShift[i]);
       if (lastScoreByShift[i] > bestScore) {
         bestScore = lastScoreByShift[i];
