### 2021-11-28 e71b2d9
refacor: introduce alias variables
These will be used in the following commits
