### 2021-11-28 103db79
refactor: use alias score variable
