### 2021-11-28 a2394b1
refactor: extract best alignment search to a function
