### 2021-11-28 dbd6ba0
refactor: inline single-use variable
