### 2021-11-28 bb64451
refactor: move shift  variable into the function
