### 2021-11-28 4200a22
refactor: reduce scope of the origin variable
