### 2021-11-28 56a6e48
refactor: move qPos calculation into the function
