### 2021-11-28 b995701
refactor: inline bestScore variable
