### 2021-11-28 d6b06a2
refactor: use variable instead of .size() call
