### 2021-11-28 9be6f27
fix: use correct size variable
Bug introduced during refactoring in e653ad1e36135f6049eace5f5f01c648448d95fc
