### 2021-11-29 a2b734e
chore: release cli 1.5.1 and web 1.8.1
