### 2021-11-30 4ca2086
refactor: remove intel tbb, move nextalign tbb filters to class methods
