+++ packages/nextalign_cli/src/cli.cpp
@@ -1,18 +1,18 @@
 #include <fmt/format.h>
 #include <nextalign/nextalign.h>
-#include <tbb/global_control.h>
-#include <tbb/parallel_pipeline.h>
 
 #include <boost/algorithm/string.hpp>
 #include <boost/algorithm/string/join.hpp>
 #include <boost/algorithm/string/split.hpp>
 #include <cxxopts.hpp>
 #include <fstream>
+#include <thread>
 
 #include "Logger.h"
 #include "description.h"
 #include "filesystem.h"
 #include "malloc_conf.h"
+#include "utils/safe_cast.h"
 
 struct CliParams {
   int jobs{};
@@ -655,127 +655,145 @@ void writeReference(                                           //
   }
 }
 
-/**
- * Runs nextalign algorithm in a parallel pipeline
- */
-void run(
-  /* in  */ int parallelism,
-  /* in  */ bool inOrder,
-  /* inout */ std::unique_ptr<FastaStream> &inputFastaStream,
-  /* in  */ const ReferenceSequenceData &refData,
-  /* in  */ const GeneMap &geneMap,
-  /* in  */ const NextalignOptions &options,
-  /* out */ std::ostream &outputFastaStream,
-  /* out */ std::ostream &outputInsertionsStream,
-  /* out */ std::ostream &outputErrorsFile,
-  /* out */ std::map<std::string, std::ofstream> &outputGeneStreams,
-  /* in */ bool shouldWriteReference,
-  /* out */ Logger &logger) {
-  tbb::task_group_context context;
-  const auto ioFiltersMode = inOrder ? tbb::filter_mode::serial_in_order : tbb::filter_mode::serial_out_of_order;
+class Nextalign {
+  std::unique_ptr<FastaStream> &inputFastaStream;
+  const GeneMap &geneMap;
+  const NextalignOptions &options;
+  std::ostream &outputFastaStream;
+  std::ostream &outputInsertionsStream;
+  std::ostream &outputErrorsFile;
+  std::map<std::string, std::ofstream> &outputGeneStreams;
+  Logger &logger;
+  const NucleotideSequence ref;
+  const std::string refName;
+  std::map<std::string, RefPeptideInternal> refPeptides;
 
-  const auto &ref = refData.seq;
-  const auto &refName = refData.name;
-
-  const auto refPeptides = translateGenesRef(ref, geneMap, options);
-
-  if (shouldWriteReference) {
-    writeReference(refName, ref, refPeptides, outputFastaStream, outputGeneStreams);
+  /** Reads contents of the input stream, parses it and and returns next parsed sequence */
+  std::optional<AlgorithmInput> readInput() {
+    if (!inputFastaStream->good()) {
+      return std::optional<AlgorithmInput>{};
+    }
+    return inputFastaStream->next();
   }
 
-  /** Input filter is a serial input filter function, which accepts an input stream,
-   * reads and parses the contents of it, and returns parsed sequences */
-  const auto inputFilter = tbb::make_filter<void, AlgorithmInput>(ioFiltersMode,//
-    [&inputFastaStream](tbb::flow_control &fc) -> AlgorithmInput {
-      if (!inputFastaStream->good()) {
-        fc.stop();
-        return {};
-      }
+  /** Runs algorithm for 1 sequence */
+  AlgorithmOutput runAlgorithm(const AlgorithmInput &input) {
+    try {
+      const auto query = toNucleotideSequence(input.seq);
+      const auto result = nextalign(query, ref, refPeptides, geneMap, options);
+      return {.index = input.index, .seqName = input.seqName, .hasError = false, .result = result, .error = nullptr};
+    } catch (const std::exception &e) {
+      const auto &error = std::current_exception();
[…]
