### 2021-11-30 55dd6da
perf: reimplement producer-consumer parallel pipeline
