+++ packages/nextalign_cli/src/cli.cpp
@@ -1,18 +1,25 @@
+#include <ThreadPool/ThreadPool.h>
 #include <fmt/format.h>
 #include <nextalign/nextalign.h>
+#include <rigtorp/MPMCQueue.h>
 
 #include <boost/algorithm/string.hpp>
 #include <boost/algorithm/string/join.hpp>
 #include <boost/algorithm/string/split.hpp>
+#include <condition_variable>
 #include <cxxopts.hpp>
 #include <fstream>
+#include <future>
+#include <mutex>
 #include <thread>
+#include <vector>
 
 #include "Logger.h"
 #include "description.h"
 #include "filesystem.h"
 #include "malloc_conf.h"
 #include "utils/safe_cast.h"
+
 
 struct CliParams {
   int jobs{};
@@ -784,13 +791,153 @@ public:
   }
 
   void runConcurrently(int parallelism, bool inOrder) {
-    // TODO: implement concurrency
-    std::optional<AlgorithmInput> input = readInput();
-    while (input) {
-      const auto output = runAlgorithm(*input);
-      writeOutput(output);
-      input = readInput();
+    constexpr int INPUT_QUEUE_SIZE = 10;
+    constexpr int OUTPUT_QUEUE_SIZE = 10;
+
+    const int numWorkers = parallelism;
+
+
+    bool isReaderThreadDone = false;
+    bool areAlgorithmThreadsDone = false;
+
+    std::mutex inputMutex;
+    std::condition_variable inputCond;
+    rigtorp::mpmc::Queue<AlgorithmInput> inputQueue(INPUT_QUEUE_SIZE);
+
+    std::mutex outputMutex;
+    std::condition_variable outputCond;
+    rigtorp::mpmc::Queue<AlgorithmOutput> outputQueue(OUTPUT_QUEUE_SIZE);
+
+    // Schedule 1 input reader
+    auto readerThread = std::thread([this, &inputQueue, &inputCond]() {
+      // fmt::print("reader start\n");
+      std::optional<AlgorithmInput> input = readInput();
+      int i = 0;
+      while (input) {
+        // Attempt to push the newly produced input and block if the input queue is full, waiting until there's room
+        // fmt::print("reader push: {:2d}\n", i);
+        inputQueue.emplace(std::move(*input));
+        inputCond.notify_one();
+        input = readInput();
+        ++i;
+      }
+      // fmt::print("reader done\n");
+    });
+
+
+    // Schedule desired number of algorithm workers
+    std::vector<std::thread> algorithmThreads;
+    algorithmThreads.reserve(numWorkers);
+    for (int i = 0; i < numWorkers; ++i) {
+      algorithmThreads.emplace_back(
+        std::thread([this, &inputQueue, &outputQueue, &isReaderThreadDone, i, &inputMutex, &inputCond, &outputCond]() {
+          // fmt::print("worker {:2d} start\n", i);
+
+          // While reader is not done and is still pushing new items into the input queue
+          while (true) {
+            // Attempt to pop an item from input queue and block if there isn't any, waiting until new items appear
+            // fmt::print("worker {:2d} pop: start\n", i);
+
+            AlgorithmInput input;
+            bool hasItem = false;
+
+            {
+              std::unique_lock<std::mutex> lock(inputMutex);
+              inputCond.wait(lock, [&hasItem, &inputQueue, &input, &isReaderThreadDone]() {
+                hasItem = inputQueue.try_pop(input);
+                auto shouldContinue = hasItem || isReaderThreadDone;
+                return shouldContinue;
+              });
+            }
+
+            if (!hasItem) {
+              break;
+            }
+
+            // fmt::print("worker {:2d} pop: ok\n", i);
+
+            auto output = runAlgorithm(input);
+
[…]
