### 2021-11-30 bffec1d
perf: try alternative fasta parser
