+++ packages/nextalign/src/io/kseq.h
@@ -1,0 +1,289 @@
+/**
+ * Taken with modifications from
+ * https://github.com/lh3/readfq/blob/091bc699beee3013491268890cc3a7cbf995435b/kseq.h
+ *
+ */
+
+/* The MIT License
+
+   Copyright (c) 2008, 2009, 2011 Attractive Chaos <attractor@live.co.uk>
+
+   Permission is hereby granted, free of charge, to any person obtaining
+   a copy of this software and associated documentation files (the
+   "Software"), to deal in the Software without restriction, including
+   without limitation the rights to use, copy, modify, merge, publish,
+   distribute, sublicense, and/or sell copies of the Software, and to
+   permit persons to whom the Software is furnished to do so, subject to
+   the following conditions:
+
+   The above copyright notice and this permission notice shall be
+   included in all copies or substantial portions of the Software.
+
+   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+   EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+   MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+   NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+   BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+   ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+   CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+   SOFTWARE.
+*/
+
+/* Last Modified: 05MAR2012 */
+
+#ifndef AC_KSEQ_H
+#define AC_KSEQ_H
+
+#include <ctype.h>
+#include <stdlib.h>
+#include <string.h>
+
+#define KS_SEP_SPACE 0// isspace(): \t, \n, \v, \f, \r
+#define KS_SEP_TAB 1  // isspace() && !' '
+#define KS_SEP_LINE 2 // line separator: "\n" (Unix) or "\r\n" (Windows)
+#define KS_SEP_MAX 2
+
+#ifndef klib_unused
+#if (defined __clang__ && __clang_major__ >= 3) || (defined __GNUC__ && __GNUC__ >= 3)
+#define klib_unused __attribute__((__unused__))
+#else
+#define klib_unused
+#endif
+#endif /* klib_unused */
+
+#define __KS_TYPE(type_t)         \
+  typedef struct __kstream_t {    \
+    int begin, end;               \
+    int is_eof : 2, bufsize : 30; \
+    type_t f;                     \
+    unsigned char *buf;           \
+  } kstream_t;
+
+#define ks_eof(ks) ((ks)->is_eof && (ks)->begin >= (ks)->end)
+#define ks_rewind(ks) ((ks)->is_eof = (ks)->begin = (ks)->end = 0)
+
+#define __KS_BASIC(SCOPE, type_t, __bufsize)                    \
+  SCOPE kstream_t *ks_init(type_t f) {                          \
+    kstream_t *ks = (kstream_t *) calloc(1, sizeof(kstream_t)); \
+    ks->f = f;                                                  \
+    ks->bufsize = __bufsize;                                    \
+    ks->buf = (unsigned char *) malloc(__bufsize);              \
+    return ks;                                                  \
+  }                                                             \
+  SCOPE void ks_destroy(kstream_t *ks) {                        \
+    if (!ks)                                                    \
+      return;                                                   \
+    free(ks->buf);                                              \
+    free(ks);                                                   \
+  }
+
+#define __KS_INLINED(__read)                                                               \
[…]
